% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{add_new_cells}
\alias{add_new_cells}
\title{Add new data}
\usage{
add_new_cells(
  new_raw_data_csv,
  cell_characteristics_csv,
  old_raw_data_csv,
  data_type,
  software = "MiniAnalysis",
  bin_width = 5,
  write_new_csv = "yes",
  new_file_name,
  decimal_places = 2,
  injection_start_time = 265.4,
  length_of_current_injection = 0.5
)
}
\arguments{
\item{new_raw_data_csv}{A filepath to a csv containing the new raw data. If
the data are evoked current data (\code{current_type == "eEPSC"}) then this must
contain 4 columns: \code{letter}, \code{ID}, \code{P1} and \code{P2}. If the data are
spontaneous current data, the columns must be \code{letter}, \code{ID},
\code{recording_num}, \code{trace}, \code{amplitude} and \code{time_of_peak}. Please see
the section below on required columns for more details.}

\item{cell_characteristics_csv}{A filepath to a csv containing information
about the cells. Please see \code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}} for a
description of what columns should be included. Don't forget to update this
to include the cell characteristics for the new letters in
\code{new_raw_data_csv}!}

\item{old_raw_data_csv}{A filepath to a csv containing the current raw data.
Since this function appends the new data to the old data, this must be of
the same current_type as the new data (e.g. the columns must be the same).
If this is the first time you are running this function, start with a blank
.csv file containing just some text (e.g. \code{letter}) in cell \code{A1}. This is required because R cannot recognize an empty \code{.csv} sheet as a valid file.}

\item{data_type}{A character (\code{"eEPSC"}, \code{"sEPSC"}, \code{"AP_parameter"} or \code{"AP_count"}) describing the data type that is being imported.}

\item{software}{A character (\code{"Clampfit"} or \code{"MiniAnalysis"}) describing what software tool was used to analyze the data in \code{new_raw_data_csv}. This is relevant when \code{data_type} is \code{"sEPSC"} because the exported data is different. Defaults to \code{"Clampfit"}.}

\item{bin_width}{A numeric value (defaults to \code{5}) describing the time interval in seconds (bin width) of the histogram used in MiniAnalysis. This is only relevant when \code{data_type = "sEPSC"} and \code{software = "MiniAnalysis"}.}

\item{write_new_csv}{A character (\code{"yes"} or \code{"no"}) describing if the new data
should be written to a csv file. Defaults to \code{"yes"}. Please specify
a filename for the new csv file in \code{new_file_name}.}

\item{new_file_name}{A filename for the csv containing the new data appended
to the old data. Must be a character representing a filepath to a csv file.
Examples include \code{"Data/20241118-Raw-eEPSC-data.csv"}.}

\item{decimal_places}{A numeric value indicating the number of decimal places the data should be rounded to. Used to reduce file size and prevent an incorrect representation of the number of significant digits.}

\item{injection_start_time}{For action potential data only: A numeric value describing the start time (in ms) when current injection was applied. Used to calculate the latency to fire.}

\item{length_of_current_injection}{For action potential data only: A numeric value indicating the duration of the current injection (in s, default is 0.5 s).}
}
\value{
A dataframe consisting of the old raw data with information from the new
cells appended to it. If \code{data_type = "AP_parameter"} two new columns will also be
added based on calculations from the existing columns. These are
\code{latency_to_fire} (which is \code{time_to_peak} - \code{injection_start_time}) and
\code{antipeak_time_relative_to_threshold} (which is \code{time_of_antipeak} -
\code{time_of_threshold}). If \code{data_type = "AP_count"} one new column will be added. These is \code{AP_frequency} (in Hz).
}
\description{
This function enables you to append new raw recording data onto an existing
datasheet. It makes it easy and convenient to merge the cell parameters (age,
sex, etc.) with new data and add it to your current raw data. This function
also formats the dataset so it is immediately ready for use in functions like
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.
}
\section{Required Columns}{


If the data are evoked currents (\code{data_type = "eEPSC"}), the data must
contain the following four columns:

\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{P1} A numeric value representing the
amplitude of the first evoked current in pA.
\item \code{P2} A numeric value representing the
amplitude of the second evoked current in pA.
}

If the data are spontaneous currents (\code{data_type = "sEPSC"}) and these were exported from Clampfit (\code{software = "Clampfit"}), the data
must contain the following columns:
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{recording_num} A numeric value representing the recording number.
This was incorporated before we switched to concatenating all recordings
into one, but it needs to remain here to prevent breaking previous projects.
It should be set to 1.
\item \code{trace} A numeric value representing the trace (automatically
generated in Clampfit) where the current occurred.
\item \code{amplitude} A numeric value representing the amplitude of the evoked
current in pA (automatically generated in Clampfit).
\item \code{time_of_peak} A numeric value representing the time of the peak in
milliseconds relative to trace number (automatically generated in Clampfit).
}

If the data are spontaneous currents (\code{data_type = "sEPSC"}) and these were exported from the histogram tool in MiniAnalysis (\code{software = "MiniAnalysis"}), the data
must contain the following columns:
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{amplitude} The average sEPSC amplitude, averaged over the time interval (bin width) that you chose for the histogram (automatically generated from histogram tool in MiniAnalysis).
\item \code{SE} The standard error of \code{amplitude} (automatically generated from histogram tool in MiniAnalysis).
\item \code{time} The time in seconds (exported automatically from the histogram, depending on the bin width that you set). \code{add_new_cells()} will automatically convert this to minutes.
\item \code{num_events} The number of synaptic events (sEPSCs) during the time interval (bin width). This is also automatically generated in MiniAnalysis.
}

If the data are action potential parameters (\code{data_type == "AP_parameter"}), the data
must contain the following columns:
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{state} A character (\code{"Baseline"} or \code{"Insulin"}) representing
the timepoint that the data point belongs to.
\item \code{time_of_threshold} The time (in ms) when the membrane potential
reaches the threshold value.
\item \code{threshold} The threshold (in mV). Determined using the first
derivative method, where the threshold is the membrane potential which
results in a derivative of 10 mV/ms or greater (Farries et al., 2010).
\item \code{t_11} The value of the first derivative (action potential velocity in
mV/ms) at threshold.
\item \code{first_sweep_with_APs} The sweep number of the first sweep (going from
lowest to higher current injection values) that resulted in an action
potential.
\item \code{trace_start} An automated value in Clampfit that is not used in the
analysis.
\item \code{peak_amplitude} The peak amplitude (in pA) of the action
potential relative to threshold.
\item \code{time_to_peak} The time to peak
amplitude (in ms) relative to the time of threshold.
\item \code{antipeak_amplitude} The after-hyperpolarization amplitude (in pA) relative
to threshold.
\item \code{time_of_antipeak} The time of the after-hyperpolarization (in ms).
\item \code{half_width} The half-width, which is the width of the action potential and half of the peak amplitude.
}
}

\examples{


# NOTE: Remember, if you are running this for the first time,
# don't start with a blank csv for `old_raw_data_csv`.
# `old_raw_data_csv` must have at least some text
# in cell `A1` for R to recognize it.

# NOTE: If you are importing spontaneous current data from MiniAnalysis,
# don't forget to set `bin_width` to match what you used
# in the histogram tool in MiniAnalysis!

\dontrun{
add_new_cells(
  new_raw_data_csv = "sample_new_eEPSC_data.csv",
  cell_characteristics_csv = "sample_cell_characteristics.csv",
  old_raw_data_csv = "sample_eEPSC_data.csv",
  data_type = "eEPSC",
  write_new_csv = "no",
  new_file_name = "Raw-eEPSC-Data.csv",
  decimal_places = 2
)
}

}
\seealso{
\code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}} for a list of required columns in the \code{cell_characteristics_csv}.

\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} for the next step in the analysis process.
}
