% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{get_fig_height}
\alias{get_fig_height}
\title{Get figure height}
\usage{
get_fig_height(data, plot_category, plot_treatment, plot_sex)
}
\arguments{
\item{data}{A dataframe containing the raw evoked current data generated from
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. If \code{pruned = "yes"} you must use the \verb{$individual_cells} component of a pruned dataset.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. \code{"Control"}). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{plot_sex}{A character value ("Male" or "Female") corresponding to the sex you would like to plot.}
}
\value{
A dataframe with a numerical value indicating the ideal figure height for each combination of category, treatment and sex.
}
\description{
This function produces a dataframe with the ideal figure height for the facet plots produced by \code{make_facet_plot()}. These facet plots have a three-column layout with varying numbers of rows. This function first determines the number of rows required for a three-column layout, and then determines the required figure height. Typically, this is four times the number of rows.
}
\details{
This function is ideal if combined with the chunk options of an RMarkdown document. Set \code{fig.width = 14} and \code{fig.height = get_fig_height(...)}. Replace the \code{...} with the correct category, treatment and sex you'd like to plot. Plot only one facet plot per chunk.
}
\examples{

get_fig_height(
  data = sample_raw_eEPSC_df,
  plot_category = 2,
  plot_treatment = "Control",
  plot_sex = "Male"
)

}
\seealso{
\code{\link[=make_facet_plot]{make_facet_plot()}}
}
