% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print Hello World}
\usage{
hello(name)
}
\arguments{
\item{name}{A character representing the person's name. This must be:
\itemize{
\item short
\item sweet
\item memorable
}}
}
\value{
A string
}
\description{
\code{make_normalized_EPSC_df()} creates a dataframe of evoked or spontaneous current data from a raw .csv file.
}
\details{
The function will create a new column containing the evoked/spontaneous current amplitudes
normalized relative to the mean baseline amplitude.
For evoked current data, the function also adds a column for the paired pulse ratio (P2/P1, where P2 is the amplitude of the second evoked current)
}
\section{Required columns}{

If the data are evoked currents (current_type == "eEPSC"), the .csv file must contain the following columns and data types:
\itemize{
\item \code{letter}:  A character value that is a unique identifier for a single recording. Used to link data sets for evoked or spontaneous currents and cell-characteristics.
\item \code{synapses}  A character value (e.g. "Glutamate" or "GABA").
\item \code{sex}   A character value (e.g. "Male" or "Female").
\item \code{treatment} Tab A character value (e.g. "Control", "HNMPA").
}
}

\examples{

hello("Christelinda")

}
