% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{import_ext_data}
\alias{import_ext_data}
\title{Get path to external file}
\usage{
import_ext_data(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
Access sample files in \code{inst/extdata}
directory.
}
\examples{
import_ext_data()
import_ext_data("sample_cell_characteristics.csv")
}
