% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{import_theme_colours}
\alias{import_theme_colours}
\title{Import colour theme}
\usage{
import_theme_colours(filename)
}
\arguments{
\item{filename}{A filepath to a .csv file containing colours and theme parameters. Must contain two columns:

\itemize{
\item \code{option} The name of the parameter. Required parameters include:
\itemize{
\item ``

}

\code{line_col}, \code{baseline_colour}, etc.
\item \code{value} The value of the parameter
}}
}
\value{
A dataframe with one column and row names.
}
\description{
Import colour theme
}
\examples{

import_theme_colours(import_ext_data("sample_theme_options.csv"))

}
