% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-raw-data.R
\name{load_ext_data}
\alias{load_ext_data}
\title{Get path to external file}
\usage{
load_ext_data(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
Access sample files in \code{inst/extdata}
directory.
}
\examples{
load_ext_data()
load_ext_data("sample_cell_characteristics.csv")
}
