% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{make_cv_plot}
\alias{make_cv_plot}
\title{Make a plot of coefficient of variation over time}
\usage{
make_cv_plot(
  data,
  plot_treatment = "Control",
  treatment_colour_theme,
  theme_options,
  save_plot_png = "no"
)
}
\arguments{
\item{data}{A dataframe of the pruned current data for all cells. This is the
third dataframe in the list generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control"). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"} it will also generate a
.png file in the folder \code{Figures/Evoked-currents/CV} relative to the
project directory. The treatment will be included with the filename.
}
\description{
\code{make_cv_plot()} enables you to save a plot of the coefficient of variation in evoked current amplitudes over time.
}
\examples{
make_cv_plot(
  data = sample_pruned_eEPSC_df$all_cells,
  plot_treatment = "Control",
  theme_options = sample_theme_options,
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_plot_png = "no"
)

}
\seealso{
\code{\link[=make_variance_comparison_plot]{make_variance_comparison_plot()}} to make plots of inverse
coefficient of variation squared and VMR, which are useful to determine if
a mechanism is pre- or post-synaptic.
}
