% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{make_normalized_EPSC_data}
\alias{make_normalized_EPSC_data}
\title{Import and normalize raw current data}
\usage{
make_normalized_EPSC_data(
  filename = "Data/Sample-eEPSC-data.csv",
  current_type = "eEPSC",
  min_time_value = 0,
  max_time_value = 25,
  baseline_length = 5,
  interval_length = 5,
  negative_transform_currents = "yes",
  save_output_as_RDS = "no"
)
}
\arguments{
\item{filename}{A filepath to a .csv file. See add_new_cells() for the
function that will merge raw data (a .csv with 4 columns: \code{letter}, \code{ID},
\code{P1}, and \code{P2}) and a \code{cell-characteristics.csv} file (with columns for
factors like \code{animal}, \code{age}, \code{sex}, \code{synapses}). Please see the section on
"Required columns" below.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{min_time_value}{Minimum time value (numeric; in minutes), which
defaults to 0.}

\item{max_time_value}{Maximum recording length (numeric; in minutes). All
data points will be filtered to time values less than or equal to this
value. Defaults to 25.}

\item{baseline_length}{Length of the baseline (numeric; in minutes). Refers
to data collected before applying a hormone, antagonist, or a protocol like
high frequency stimulation. Defaults to 5.}

\item{interval_length}{Length of each interval (numeric; in minutes). Used to
divide the dataset into broad ranges for statistical analysis. Important!
\code{max_recording_length} must be evenly divisible by \code{interval_length}.
Defaults to 5.}

\item{negative_transform_currents}{A character ("yes" or "no") describing if
\code{P1} and \code{P2} should be negative transformed. If "yes", the values will be
multiplied by (-1).}

\item{save_output_as_RDS}{A character ("yes" or "no") describing if the
resulting object should be saved as an RDS file in the raw data folder.}
}
\value{
A dataframe that can be viewed and used for further analyses in R.
New or modified columns include:
\itemize{
\item \code{P1} (for evoked currents only) May be negative-transformed if
\code{negative_transform} == "yes"
\item \code{P2} (for evoked currents only) May be negative-transformed if
\code{negative_transform} == "yes"
\item \code{PPR} (for evoked currents only) A numeric value that represents the
paired pulse ratio (PPR) of the evoked currents, generated using
\code{dplyr::mutate(PPR = P2/P1)}.
\item \code{interval} A character value indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \code{baseline_range} A logical value required for the baseline
transformation. It is set to TRUE when time is within the baseline period
(e.g. Time <= 5) and FALSE at all other times.
\item \code{baseline_mean} A numeric value representing the mean evoked current
amplitude during the baseline period. There is a different baseline_mean for
each letter.
\item \code{P1_transformed} (for evoked currents only) A numeric value
representing the first evoked current
amplitude (pA) normalized relative to the mean amplitude during the
recording's baseline.
\item \code{P2_transformed} (for evoked currents only) A numeric value
representing the second evoked current amplitude (pA) normalized relative to
the mean amplitude during the recording's baseline.
\item \code{amplitude_transformed} (for spontaneous currents only) A numeric
value representing the spontaneous current amplitude (pA) normalized
relative to the mean amplitude during the recording's baseline.
}
}
\description{
\code{make_normalized_EPSC_df()} creates a dataframe of evoked or spontaneous
current data from a raw .csv file. The function will create a new column
containing the evoked or spontaneous current amplitudes normalized relative
to the mean current amplitude during the baseline period. For evoked current
data, the function also adds a column for the paired pulse ratio (\code{PPR = P2/P1}, where \code{P2} is the amplitude of the second evoked current).
}
\section{Required basic columns}{

It doesn't matter if the data are evoked or current type - these columns
should be included in your data.
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{synapses} A character value (e.g. "Glutamate" or "GABA").
\item \code{sex} A character value (e.g. "Male" or "Female").
\item \code{treatment} A character value (e.g. "Control", "HNMPA").
\item \code{time} A numeric value that represents time in minutes. This column is
autogenerated in add_new_cells().
\item \code{ID} A character value for the recording filename.
\item \code{X} A numeric value representing the x-value of the cell's location in
µm. Leave this blank if you don't have this data.
\item \code{Y} A numeric value representing the y-value of the cell's location in
µm. Leave this blank if you don't have this data.
\item \code{age} A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages).
\item \code{animal} A numeric value representing the animal's ID or number.
\item \code{category} A numeric value representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups.
\item \code{cell} A character or numeric value representing the cell. For
example, use \verb{3.1.1} for animal #3, slice #1, cell #1.
}

\strong{Evoked current data}:

If the data are evoked currents (\code{current_type == "eEPSC"}), the data must
contain the basic columns mentioned in \strong{Required basic columns} plus these
columns:

\itemize{
\item \code{P1} A numeric value representing the
amplitude of the first evoked current in pA.
\item \code{P2} A numeric value representing the
amplitude of the second evoked current in pA.
}

\strong{Spontaneous current data:}

If the data are spontaneous currents (\code{current_type == "sEPSC"}), the data
must contain the basic columns mentioned in \strong{Required basic columns} plus
these columns:
\itemize{
\item \code{recording_num} A numeric value representing the recording number.
This was incorporated before we switched to concatenating all recordings
into one, but it needs to remain here to prevent breaking previous projects.
It should be set to 1.
\item \code{trace} A numeric value representing the trace (automatically
generated in Clampfit) where the current occured.
\item \code{time_of_peak} A numeric value representing the time of the peak in
milliseconds relative to trace number. This is automatically calculated in
Clampfit.
\item \code{time} A numeric value representing the absolute time when the current
happened, relative to the start of the recording. This is autogenerated. See
add_new_cells() for a description of how the true time value (\code{time}) is
calculated from the \code{recording_num} and \code{trace.}
\item \code{amplitude} A numeric value representing the amplitude of the evoked
current in pA.
}
}

\examples{
make_normalized_EPSC_data(
  filename = import_ext_data("sample_eEPSC_data.csv"),
  current_type = "eEPSC",
  min_time_value = 0,
  max_time_value = 25,
  interval_length = 5,
  baseline_length = 5,
  negative_transform_currents = "yes"
)

}
\seealso{
add_new_cells() to add new recording data to your existing raw
csv. It will merge raw data (a .csv with 4 columns: \code{letter}, \code{ID}, \code{P1},
and \code{P2}) and a \code{cell-characteristics.csv} file (with columns for factors
like \code{animal}, \code{age}, \code{sex}, \code{synapses}).
}
