% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{make_PPR_data}
\alias{make_PPR_data}
\title{Make paired-pulse ratio (PPR) dataframe for before vs. after comparisons}
\usage{
make_PPR_data(
  data,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  PPR_min = 0,
  PPR_max = 5,
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  treatment_colour_theme,
  save_output_as_RDS = "no"
)
}
\arguments{
\item{data}{A dataframe containing the raw evoked current data generated from
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{include_all_treatments}{A character (\code{"yes"} or \code{"no"}) specifying if the
plot will include data from all treatments. If \code{"no"}, you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to \code{NULL}, since
include_all_treatments is \code{"yes"} by default.}

\item{PPR_min}{A numeric value representing the minimum PPR value permitted
in the filtered dataset. Defaults to \code{0}.}

\item{PPR_max}{A numeric value representing the maximum PPR value permitted
in the filtered dataset. Defaults to \code{5}.}

\item{baseline_interval}{A character value indicating the name of the
interval used as the baseline. Defaults to \code{"t0to5"}, but can be changed.
Make sure that this matches the baseline interval that you set in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{post_hormone_interval}{A character value specifying the interval used
for the data points after a hormone or protocol was applied.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_output_as_RDS}{A character (\code{"yes"} or \code{"no"}) describing if the
resulting object should be saved as an RDS file in the raw data folder.}
}
\value{
A dataframe containing all of the columns from
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}, but filtered to only include PPR values
between \code{PPR_min} and \code{PPR_max} within the \code{baseline_interval} and
\code{post_hormone_interval}.
}
\description{
This function filters the raw current data into data that belong to one of
two time points. They are the baseline period and a user-specified interval
after a hormone or protocol has been applied. The "before/after" comparison
of the paired-pulse ratio (PPR) is useful to determine which mechanism is
involved in modifying synaptic plasticity. For example, the PPR may be
related to the probability of neurotransmitter release \href{https://doi.org/10.1111/j.1469-7793.2000.00513.x}{Oleskevich et al., 2000}.
}
\examples{

make_PPR_data(
  data = sample_raw_eEPSC_df,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  PPR_min = 0,
  PPR_max = 5,
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  treatment_colour_theme = sample_treatment_names_and_colours
)

}
\seealso{
\code{\link[=plot_PPR_data_single_treatment]{plot_PPR_data_single_treatment()}} and \code{\link[=plot_PPR_data_multiple_treatments]{plot_PPR_data_multiple_treatments()}}
}
