% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{make_raw_plots}
\alias{make_raw_plots}
\title{Make raw current plots}
\usage{
make_raw_plots(
  data,
  plot_treatment,
  plot_category,
  current_type,
  parameter,
  pruned,
  hormone_added,
  hormone_or_HFS_start_time,
  hormone_end_time = NULL,
  save_plot_png = "no",
  theme_options,
  treatment_colour_theme
)
}
\arguments{
\item{data}{A dataframe containing the raw evoked current data generated from
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control").}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{parameter}{A character value specifying the parameter to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
parameters are "P1", "mean_P1" and "PPR". \emph{Note}: If you select "mean_P1",
you must set the \code{pruned} argument to "yes". For spontaneous currents
(\code{current_type = "sEPSC"}), the available parameters are "amplitude" or
"frequency".}

\item{pruned}{A character value ("yes" or "no") specifying if the data are
pruned. This is only used for evoked current data where \verb{parameter = "mean_P1}. The plot will then present the data as means with error bars.}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
"500 nM Insulin", "CCK + Leptin", and "Insulin". If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{hormone_or_HFS_start_time}{A numeric value indicating the time (in
minutes) when a hormone was added or when HFS was applied. This will set
the annotation line start point.}

\item{hormone_end_time}{A numeric value indicating the time (in minutes) when
a hormone stopped being applied, such as for a washout experiment.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}
}
\value{
A list of ggplot objects, where each list element is a scatterplot of
one recording.
}
\description{
This function enables you to create a scatterplot of raw evoked (or
spontaneous) current amplitude over time in minutes. The plot title will be
pulled from the recording ID (the \code{letter} column in the raw data), and the
subtitle will include the sex and treatment. The plot will also contain a
horizontal line annotation displayed over the time region where a hormone was
applied. If a high-frequency stimulation (HFS) protocol is used, the plot
will display an arrow with the label "HFS" to indicate the time when HFS was
applied.
}
\examples{

make_raw_plots(
 data = sample_raw_eEPSC_df,
 plot_treatment = "Control",
 plot_category = 2,
 current_type = "eEPSC",
 parameter = "P1",
 pruned = "no",
 hormone_added = "Insulin",
 hormone_or_HFS_start_time = 5,
 theme_options = sample_theme_options,
 treatment_colour_theme = sample_treatment_names_and_colours
)


}
