% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{make_summary_EPSC_data}
\alias{make_summary_EPSC_data}
\title{Summarize current data per 5-min for statistical tests}
\usage{
make_summary_EPSC_data(
  data = patchclampplotteR::sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no",
  decimal_places = 2,
  baseline_interval = "t0to5",
  ending_interval = "t20to25"
)
}
\arguments{
\item{data}{A \code{data.frame} object. If the data are evoked currents
(\code{current_type == "eEPSC"}), this should be the raw evoked current data
generated using \code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. If the data are spontaneous
currents (\code{current_type == "sEPSC"}), this should be the pruned data
\verb{$individual_cells} dataset generated using \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{save_output_as_RDS}{A character ("yes" or "no") describing if the
resulting object should be saved as an RDS file in the raw data folder.}

\item{decimal_places}{A numeric value indicating the number of decimal places
the data should be rounded to. Used to reduce file size and prevent an
incorrect representation of the number of significant digits.}

\item{baseline_interval}{A character value describing the baseline interval. Defaults to "t0to5".}

\item{ending_interval}{A character value describing the last interval in the recording. Useful for future plots in which you compare the percent decrease/increase in current amplitude relative to the baseline. Examples include "t20to25", "t10to15", etc.}
}
\value{
A list of two dataframes (\code{current_type = eEPSC}) or three dataframes (\code{current_type = sEPSC}). For evoked currents (\code{current_type = "eEPSC"}) the first dataframe (\verb{$percent_change_data}) contains the mean current amplitude for each interval, with a final column (\code{percent_change}) containing the final percent change in amplitude in the last interval relative to the mean amplitude during the baseline interval.

Most columns (\code{age}, \code{sex}, \code{animal}, etc.) come directly from the information imported through \code{import_cell_characteristics_df()}. However, there are some new columns of note.

\itemize{
\item \code{t0to5} The mean evoked current amplitude (pA) for this cell during the period of 0 to 5 minutes.
\item \code{t5to10} The mean evoked current amplitude (pA) for this cell during the period of 5 to 10 minutes.
\item \code{t10to15}, \code{t15to20}, \code{tXtY} etc... The mean evoked current amplitude (pA) for this cell during the period of \emph{X} to \emph{Y} minutes.
\item \code{percent_change} The percent change in evoked current amplitude in the interval \code{t20to25} as a percentage of the mean baseline amplitude (\code{t0to5}). For example, if currents began at 100 pA during the baseline period, but were 50 pA by \code{t20to25}, the value of \code{percent_change} will be \verb{50\%} or \code{0.50}. You can also change the value of the intervals used in this calculation through the \code{baseline_interval} and \code{ending_interval} arguments.
}

The second dataframe (accessed through \verb{$summary_data}) contains summary data such as the mean current amplitude,
coefficient of variation, standard deviation, standard error, variance,
variance-to-mean ratio, and inverse coefficient of variation squared for
each interval.

New columns for evoked current data (\code{current_type == "eEPSC"}) include:
\itemize{
\item \code{mean_P1_transformed} The amplitude of the first evoked current
amplitude (\% Baseline eEPSC amplitude) normalized to the mean baseline
amplitude and averaged over the interval.
\item \code{mean_P1_raw} The amplitude of the first evoked current amplitude (pA)
averaged over the interval.
\item \code{n} The number of datapoints used to create the averaged values.
Corresponds to the number of sweeps per interval.
\item \code{sd} The standard deviation of the normalized evoked current data
(\code{P1_transformed}).
\item \code{cv} The coefficient of variation of \code{P1_transformed}.
\item \code{se} The standard error of \code{P1_transformed}.
\item \code{cv_inverse_square} The inverse of the squared coefficient of
variation of \code{P1_transformed}.
\item \code{variance} The variance of \code{P1_transformed}.
\item \code{VMR} The variance-to-mean ratio (VMR) of \code{P1_transformed}.
\item \code{interval} A character value indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \verb{letter, synapses, sex, treatment, etc.} Unmodified columns from the
original dataset describing the cell's properties.
}
\subsection{Spontaneous Current Data}{

Spontaneous current data results in three dataframes. The first dataframe ($summary_data) contains summary statistics for each interval, as outlined below:

\itemize{
\item \code{mean_transformed_amplitude} The average normalized spontaneous
current amplitude (\% Baseline sEPSC amplitude).
\item \code{mean_raw_amplitude} The average raw spontaneous current amplitude
(pA).
\item \code{n} The number of datapoints used to create the average.
\item \code{sd_transformed_amplitude} The standard deviation of the normalized
spontaneous current data (\code{mean_transformed_amplitude}).
\item \code{se_transformed_amplitude} The standard error of
\code{mean_transformed_amplitude}.
\item \code{mean_transformed_frequency} The average normalized frequency (\%
Baseline frequency).
\item \code{sd_transformed_frequency} The standard deviation of \code{mean_transformed_frequency}.
\item \code{se_frequency} The standard error of \code{mean_transformed_frequency}.
\item \code{mean_raw_frequency} The average raw frequency (Hz).
\item \verb{letter, synapses, sex, treatment, etc.} Unmodified columns from the
original dataset describing the cell's properties.
}

The second and third dataframes contain percent change data for spontaneous current amplitude and frequency, respectively. The columns are the same as the ones produced for evoked currents.
}
}
\description{
\code{make_summary_EPSC_data()} allows you to divide data from a long recording
(e.g. 30 minutes) into evenly-spaced intervals (e.g. 5 minutes). It will
generate summary data like the mean current amplitude for each interval. This
can be useful for inserting into statistical models to compare effect sizes
across broad stretches of time. The interval length would have been
previously specified in \code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} using the
\code{interval_length} argument.
}
\examples{

# Evoked Currents
# Will return a list of two dataframes

make_summary_EPSC_data(
  data = sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no",
  decimal_places = 2
)

# Spontaneous Data
# Will return a list of three dataframes

make_summary_EPSC_data(
  data = sample_pruned_sEPSC_df$individual_cells,
  current_type = "sEPSC",
  save_output_as_RDS = "no",
  decimal_places = 2
)

}
\seealso{
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}} for an example of a function that
further processes the data.
}
