% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{make_summary_plot}
\alias{make_summary_plot}
\title{Make a summary plot for a specific treatment}
\usage{
make_summary_plot(
  plot_category,
  plot_treatment,
  data,
  current_type,
  parameter,
  hormone_added,
  hormone_or_HFS_start_time,
  include_representative_trace = "no",
  representative_trace_filename,
  signif_stars = "no",
  large_axis_text = "no",
  shade_intervals = "no",
  theme_options,
  treatment_colour_theme,
  save_plot_png = "no"
)
}
\arguments{
\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control").}

\item{data}{A dataframe containing the raw evoked current data generated from
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{parameter}{A character value specifying the parameter to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
parameters are "P1", "mean_P1" and "PPR". \emph{Note}: If you select "mean_P1",
you must set the \code{pruned} argument to "yes". For spontaneous currents
(\code{current_type = "sEPSC"}), the available parameters are "amplitude" or
"frequency".}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
"500 nM Insulin", "CCK + Leptin", and "Insulin". If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{hormone_or_HFS_start_time}{A numeric value indicating the time (in
minutes) when a hormone was added or when HFS was applied. This will set
the annotation line start point.}

\item{include_representative_trace}{A character ("yes" or "no") describing if a representative trace should be included as an overlay to the plot. This pulls from a png file stored in \verb{Figures/Representative-Traces/}". Please specify the file-name in \code{representative_trace_filename}.}

\item{representative_trace_filename}{A character value describing the filename of the representative trace. This should be the name of a .png file stored within the subfolder \verb{Figures/Representative-Traces/}.}

\item{signif_stars}{A character ("yes" or "no") describing if significance stars should be included as an overlay in the plot.}

\item{large_axis_text}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which increases the size of the axis text.}

\item{shade_intervals}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which adds lightly shaded rectangles to highlight 5-minute intervals.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}
}
\value{
A ggplot object
}
\description{
This function enables you to create a scatterplot of mean evoked (or
spontaneous) current amplitude over over time in minutes. The data are summarized by treatment and sex, and averaged per minute. The data are presented as mean +/- the standard error.
}
