% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{make_variance_data}
\alias{make_variance_data}
\title{Make variance analysis dataframe}
\usage{
make_variance_data(
  data,
  df_category,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  baseline_interval,
  post_hormone_interval,
  treatment_colour_theme,
  save_output_as_RDS = "no"
)
}
\arguments{
\item{data}{A dataframe containing the summary data generated from
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}.}

\item{df_category}{A numeric value describing the experimental category. In
the sample dataset for this package, 2 represents experiments where insulin
was applied continuously after a 5-minute baseline period. Here,
\code{plot_treatment} represents antagonists that were present on the brain
slice, or the animals were fasted, etc.}

\item{include_all_treatments}{A character ("yes" or "no") specifying if the
plot will include data from all treatments. If "no", you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to NULL, since
include_all_treatments is "yes" by default.}

\item{baseline_interval}{A character value indicating the name of the
interval used as the baseline. Defaults to "t0to5", but can be changed.
Make sure that this matches the baseline interval that you set in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{post_hormone_interval}{A character value indicating the name of the
interval used as "after" timepoint for comparison. Defaults to "t20to25",
but can be changed. Make sure that this matches an interval present in
\code{data}}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_output_as_RDS}{A character ("yes" or "no") describing if the
resulting object should be saved as an RDS file in the raw data folder.}
}
\value{
A dataframe containing all of the columns within the summary data
(see \link{sample_summary_eEPSC_df} for a detailed description of these columns)
plus three additional columns:
\itemize{
\item \code{state} A character value describing if a data point belongs to the
baseline interval ("Baseline") or an interval after a hormone or protocol has
been applied ("Post-modification"). These intervals are selected from
\code{baseline_interval} and \code{post_hormone_interval}.
\item \code{mean_cv_inverse_square} The mean inverse coefficient of variation squared within a specific state.
\item \code{mean_VMR} The mean variance-to-mean ratio within a specific state.
}
}
\description{
\code{make_variance_data} creates a dataframe containing variance measures at two
time points. They are the baseline period and a user-specified interval after
a hormone or protocol has been applied. The variance measures are the inverse
coefficient of variation squared and the variance-to-mean ratio (VMR). A
"before vs. after" comparison of these two variance measures is useful to
determine which mechanism is involved in modifying synaptic plasticity. For
more information, please see \href{https://doi.org/10.1016/j.jneumeth.2019.108526}{Huijstee & Kessels (2020)}.
}
\examples{

make_variance_data(data = sample_summary_eEPSC_df,
 df_category = 2,
 include_all_treatments = "yes",
 list_of_treatments = NULL,
 baseline_interval = "t0to5",
 post_hormone_interval = "t20to25",
 treatment_colour_theme = sample_treatment_names_and_colours,
 save_output_as_RDS = "no"
)

}
