% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchclampplotteR-package.R
\docType{package}
\name{patchclampplotteR-package}
\alias{patchclampplotteR}
\alias{patchclampplotteR-package}
\title{patchclampplotteR: Plot and Analyze Raw Patch Clamp Electrophysiology Data}
\description{
Create publication-quality figures from whole-cell patch clamp electrophysiology data.
}
\keyword{internal}
