% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{perform_t_tests_for_summary_plot}
\alias{perform_t_tests_for_summary_plot}
\title{Perform t-tests (or Wilcoxon tests) for EPSC summary plots}
\usage{
perform_t_tests_for_summary_plot(
  data,
  test_category,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  current_type = "eEPSC",
  parameter = "amplitude",
  baseline_interval = "t0to5",
  interval_length = 5,
  treatment_colour_theme,
  test_type = "pairwise.t.test",
  p_adjust_method = "holm",
  save_output_as_RDS = "no"
)
}
\arguments{
\item{data}{A dataframe containing the summary data generated from
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}. If \code{current_type} is "eEPSC", this must be the \verb{$summary_data} element of the list produced by \code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}.}

\item{test_category}{A numeric value describing the experimental category. In
the sample dataset for this package, 2 represents experiments where insulin
was applied continuously after a 5-minute baseline period.}

\item{include_all_treatments}{A character (\code{"yes"} or \code{"no"}) specifying if the
plot will include data from all treatments. If \code{"no"}, you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to \code{NULL}, since
include_all_treatments is \code{"yes"} by default.}

\item{current_type}{A character describing the current type. Allowed values
are \code{"eEPSC"} or \code{"sEPSC"}.}

\item{parameter}{A character value specifying the parameter to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
parameter is "amplitude", which contains amplitudes normalized relative to
the baseline. For spontaneous currents (\code{current_type = "sEPSC"}), the
available parameters are "amplitude" (normalized currents),
"raw_amplitude", "frequency" (normalized frequency) or "raw_frequency".}

\item{baseline_interval}{A character value indicating the name of the
interval used as the baseline. Defaults to \code{"t0to5"}, but can be changed.
Make sure that this matches the baseline interval that you set in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{interval_length}{Length of each interval (numeric; in minutes). Used to
divide the dataset into broad ranges for statistical analysis. Important!
\code{max_recording_length} must be evenly divisible by \code{interval_length}.
Defaults to \code{5}.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{test_type}{A character (must be \code{"pairwise.wilcox.test"} or \code{"pairwise.t.test"})
describing the statistical model used in this function.}

\item{p_adjust_method}{This argument is directly related to \code{p.adjust.method} in \code{rstatix::t_test}. This is the method used to adjust the p-value in multiple pairwise comparisons. Allowed values include \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, \code{"none"} (although \code{"none"} is not recommended).}

\item{save_output_as_RDS}{A character (\code{"yes"} or \code{"no"}) describing if the
resulting object should be saved as an RDS file in the raw data folder.}
}
\value{
A dataframe
}
\description{
This function enables you to perform a series of paired t-tests (or Wilcoxon tests) comparing the
mean current amplitude within each interval relative to the mean current
amplitude during the baseline. This uses the \code{pairwise_t_test} or \code{pairwise_wilcox_test} functions from
\code{rstatix}, with \code{paired = TRUE} and Holm's adjustment for multiple
comparisons (\code{p_adjust_method = "holm"}) by default. The resulting output table can also
be used to apply significance stars to the plot in
\code{\link[=plot_summary_current_data]{plot_summary_current_data()}}.
}
\examples{
perform_t_tests_for_summary_plot(
  data = sample_summary_eEPSC_df$summary_data,
  test_category = 2,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  current_type = "eEPSC",
  parameter = "amplitude",
  baseline_interval = "t0to5",
  test_type = "pairwise.t.test",
  interval_length = 5,
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_output_as_RDS = "no"
)
}
\references{
Nutter B (2018). \emph{lazyWeave: LaTeX Wrappers for R #' Users}. R package version
3.0.2, \url{https://CRAN.R-project.org/package=lazyWeave}.
}
\seealso{
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} for an example of how the normalized
current amplitudes were created.
}
