% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_AP_trace}
\alias{plot_AP_trace}
\title{Plot action potential recording}
\usage{
plot_AP_trace(
  file,
  sweeps,
  trace_color,
  line_width = 0.7,
  plot_category,
  plot_treatment,
  state,
  save_plot_png = "no",
  filename_suffix
)
}
\arguments{
\item{file}{A dataframe generated using \code{import_ABF_file()} with \code{recording_mode = "current_clamp"}.}

\item{sweeps}{A character value or list of character values of the sweeps you would like to plot. These correspond to the values in the \code{sweep1} column of your dataset, and will likely be in the form of "epi1", "epi2", etc.}

\item{trace_color}{A hex value of the colour of the lineplot}

\item{line_width}{A numeric value specifying the width of the lineplot}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control"). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{state}{A character value describing if the recording was taken during the baseline period or post-treatment/protocol. Examples include "Baseline", "Post-insulin". The \code{state} will be included in the .png filename if \code{save_plot_png = "yes"}.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you
have specified a custom list of treatments.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"}, it will also generate
a .png file in the folder \code{Figures/Action-potentials/Representative-traces} relative to the
project directory.
}
\description{
This function allows you to plot an \code{.abf} file of a recording taken in current clamp mode. It is useful if you want to display a representative trace of action potentials or the results of a current injection protocol.
}
\examples{

plot_AP_trace(file = sample_ap_abf_baseline,
sweeps = c("epi1", "epi10"),
trace_color = "orange",
plot_category = 2,
plot_treatment = "Control",
state = "Baseline")

}
