% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\encoding{UTF-8}
\name{plot_cell_coordinates_data}
\alias{plot_cell_coordinates_data}
\title{Plot cell location data}
\usage{
plot_cell_coordinates_data(
  data,
  background_slice_filename,
  background_slice_height = 800,
  background_slice_width = 800,
  plot_category,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  scale_bar_x_start = 700,
  scale_bar_x_length = 100,
  scale_bar_y_start = 100,
  scale_bar_y_length = 100,
  scale_bar_units = "um",
  include_scale_bar_label = "yes",
  scale_bar_thickness = 0.8,
  scale_bar_colour = "white",
  scale_bar_label_size = 4,
  geom_point_size = 3,
  geom_text_family = "",
  geom_point_alpha = 0.95,
  legend_width = 0.4,
  legend_title = "Change in eEPSC\\namplitude (\%)",
  legend_text_size = 6,
  legend_height = 10,
  filename_suffix,
  save_plot_png = "no",
  theme_options,
  ggplot_theme = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe generated from \code{make_summary_EPSC_data()}. If \code{current_type = "eEPSC"}, this must be the percent change dataframe generated from \code{make_summary_EPSC_data()}. Use \verb{$percent_change} to access this dataframe. If \code{current_type = "sEPSC"}, this can either be \verb{$percent_change_amplitude} or \verb{$percent_change_frequency}.}

\item{background_slice_filename}{A character value describing the location of the background slice image. This may be a png or a jpeg. It must contain just the background slice, and the height and width must be known values in \code{scale_bar_units}.}

\item{background_slice_height}{The height of the background slice image in \code{scale_bar_units}.}

\item{background_slice_width}{The width of the background slice image in \code{scale_bar_units}.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{include_all_treatments}{A character ("yes" or "no") specifying if the
plot will include data from all treatments. If "no", you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to NULL, since
include_all_treatments is "yes" by default.}

\item{scale_bar_x_start}{The starting location of the scale bar on the x-axis in \code{scale_bar_units}.}

\item{scale_bar_x_length}{The length of the scale bar on the x-axis in \code{scale_bar_units}.}

\item{scale_bar_y_start}{The starting location of the scale bar on the y-axis in \code{scale_bar_units}.}

\item{scale_bar_y_length}{The length of the scale bar on the y-axis in \code{scale_bar_units}.}

\item{scale_bar_units}{The units of the scale bar. Often, it will be "um".}

\item{include_scale_bar_label}{A character value ("yes" or "no") describing whether to include text labels on the scale bar. If "yes", the text annotation values will be pulled from \code{scale_bar_x_length} and \code{scale_bar_y_length}.}

\item{scale_bar_thickness}{A numeric value describing the thickness of the scale bar. Defaults to 0.8.}

\item{scale_bar_colour}{A character value (named colour of hex value) describing the colour of the scale bar. Defaults to "black".}

\item{scale_bar_label_size}{A numeric value describing the size of the scale bar annotation text. Defaults to 4.}

\item{geom_point_size}{A numeric value describing the size of the points. Defaults to 3.}

\item{geom_text_family}{A character value describing the font family used for the scale bar annotations. Defaults to "", but can be replaced with a named font. Use a package like \code{extrafont} to load system fonts into R.}

\item{geom_point_alpha}{A numeric value describing the transparency of the points. Defaults to 0.95, but can be reduced to better view the brain slice below.}

\item{legend_width}{The width of the legend colourbar. Defaults to 0.4.}

\item{legend_title}{A character value for the colourbar title. Defaults to "Change in eEPSC amplitude (\%)"}

\item{legend_text_size}{The size of the text labels used in the legend. Defaults to 6.}

\item{legend_height}{The height of the legend colourbar. Defaults to 10.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you
have specified a custom list of treatments.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}

\item{theme_options}{A dataframe containing theme options, defaults to \code{sample_theme_options}. See
\link{sample_theme_options} for an example of what this dataframe should
look like and how you can customize these values.}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_color_viridis_c()} such as \code{begin}, \code{end}, \code{option} and \code{direction}.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"}, it will also generate
a .png file in \code{Figures/Cell-coordinates-plot}. The .png filename will
contain the \code{plot_category}.
}
\description{
This function enables you to plot your cell locations on a coordinate system with an image of a brain slice in the background. The points are coloured by the percent change in evoked or spontaneous current amplitude over time.
}
\examples{

plot_cell_coordinates_data(
  data = sample_summary_eEPSC_df$percent_change_data,
  background_slice_filename = import_ext_data("DMH-brain-slice.jpg"),
  plot_category = 2,
  option = "plasma",
  theme_options = sample_theme_options
)

}
