% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_cv_data}
\alias{plot_cv_data}
\title{Make a plot of coefficient of variation over time}
\usage{
plot_cv_data(
  data,
  plot_treatment = "Control",
  plot_category,
  treatment_colour_theme,
  included_sexes = "both",
  male_label = "Male",
  female_label = "Female",
  x_label = "Time (min)",
  theme_options,
  save_plot_png = "no",
  filename_suffix = "",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{A dataframe of the pruned current data for all cells. This is the
third dataframe in the list generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. \code{"Control"}). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{included_sexes}{A character value (\code{"both"}, \code{"male"} or \code{"female"}). Useful if you want to have a plot with data from one sex only. Defaults to \code{"both"}. If you choose a single sex, the resulting plot will have \code{"-males-only"} or \code{"-females-only"} in the file name.}

\item{male_label}{A character value used to describe how males are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for male data in the \code{sex} column, and it must be consistent across data sheets. Defaults to \code{"Male"}.}

\item{female_label}{A character value used to describe how females are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for female data in the \code{sex} column, and it must be consistent across data sheets. This must be consistent in all data sheets. Defaults to \code{"Female"}.}

\item{x_label}{A character value specifying the x-axis label. Defaults to "Time (min)".}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{save_plot_png}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below \verb{Figures/} in your project directory.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you want to specify anything about the data (for example, to distinguish between recordings produced in MiniAnalysis
vs. Clampfit).}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"} it will also generate a
.png file in the folder \code{Figures/Evoked-currents/CV} relative to the
project directory. The treatment will be included with the filename.
}
\description{
\code{plot_cv_data()} enables you to save a plot of the coefficient of variation
in evoked current amplitudes over time.
}
\examples{
plot_cv_data(
  data = sample_pruned_eEPSC_df$all_cells,
  plot_treatment = "Control",
  plot_category = 2,
  theme_options = sample_theme_options,
  included_sexes = "both",
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_plot_png = "no"
)

}
\seealso{
\code{\link[=plot_variance_comparison_data]{plot_variance_comparison_data()}} to make plots of inverse
coefficient of variation squared and VMR, which are useful to determine if
a mechanism is pre- or post-synaptic.
}
