% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_PPR_data_multiple_treatments}
\alias{plot_PPR_data_multiple_treatments}
\title{Make a PPR plot for multiple treatments}
\usage{
plot_PPR_data_multiple_treatments(
  data,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  plot_category = 2,
  included_sexes = "both",
  male_label = "Male",
  female_label = "Female",
  baseline_label = "B",
  post_hormone_label = "A",
  test_type,
  map_signif_level_values = F,
  geom_signif_family = "",
  geom_signif_text_size = 3,
  geom_signif_size = 0.3,
  treatment_colour_theme,
  theme_options,
  filename_suffix = "",
  save_plot_png = "no",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{Paired pulse ratio data generated from \code{\link[=make_PPR_data]{make_PPR_data()}}.}

\item{include_all_treatments}{A character (\code{"yes"} or \code{"no"}) specifying if the
plot will include data from all treatments. If \code{"no"}, you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to \code{NULL}, since
include_all_treatments is \code{"yes"} by default.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{included_sexes}{A character value (\code{"both"}, \code{"male"} or \code{"female"}). Useful if you want to have a plot with data from one sex only. Defaults to \code{"both"}. If you choose a single sex, the resulting plot will have \code{"-males-only"} or \code{"-females-only"} in the file name.}

\item{male_label}{A character value used to describe how males are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for male data in the \code{sex} column, and it must be consistent across data sheets. Defaults to \code{"Male"}.}

\item{female_label}{A character value used to describe how females are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for female data in the \code{sex} column, and it must be consistent across data sheets. This must be consistent in all data sheets. Defaults to \code{"Female"}.}

\item{baseline_label}{A character value for the x-axis label applied to the
pre-hormone state. Defaults to \code{"Baseline"}.}

\item{post_hormone_label}{A character value for x-axis label applied to
the post-hormone or post-protocol state. Defaults to \code{"Post-hormone"} but you
will likely change this to the hormone or protocol name.}

\item{test_type}{A character (must be \code{"wilcox.test"}, \code{"t.test"} or \code{"none"})
describing the statistical model used to create a significance bracket
comparing the pre- and post-hormone groups.}

\item{map_signif_level_values}{A \code{TRUE/FALSE} value or a list of character values for mapping p-values. If \code{TRUE}, p-values will be mapped with asterisks (e.g. \* for p < 0.05, for p < 0.01). If \code{FALSE}, raw p-values will display. You can also insert a list of custom mappings or a function. For example, you can use the \code{return_p_value_as_stars()} function. If you want to write your own function, write something like \code{map_signif_level_values = function(p) if (p < 0.1) {round(p, 3)} else {"ns"}} to only display the p-values when they are below 0.1.}

\item{geom_signif_family}{A character value describing the font family used for the p-value annotations used by \code{ggsignif::geom_signif()}. Defaults to \code{""} (empty value, will be replaced with default system font), but can be replaced with a named font. Use a package like \code{extrafont} to load system fonts into R.}

\item{geom_signif_text_size}{A numeric value describing the size of the text annotations (significance stars or p-values) on the plot. Defaults to \code{8}.}

\item{geom_signif_size}{A numeric value describing the size of the \code{geom_signif} bracket size. Defaults to \code{0.3}, which is a good thickness for most applications.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{theme_options}{A dataframe containing theme options, defaults to \code{sample_theme_options}. See
\link{sample_theme_options} for an example of what this dataframe should
look like and how you can customize these values.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you
have specified a custom list of treatments.}

\item{save_plot_png}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, the plot will be
saved as a .png using \code{ggsave()}. The filepath depends on the current type, but
they will all go in subfolders below \verb{Figures/} in your project directory.}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A ggplot object. If \code{save_plot_png} is defined as "yes", it will also
generate a .png file in the folder \code{Figures/Evoked-currents/PPR} relative
to the project directory.
}
\description{
\code{plot_PPR_data_multiple_treatments()} creates a categorical scatter plot with
experimental state (i.e. grouped as baseline/before and after) and treatment
on the x-axis, and the paired-pulse ratio (PPR) on the y-axis. There are also
lines connecting the "before" data point to the "after" data point for each
letter. It is the same as \code{\link[=plot_PPR_data_single_treatment]{plot_PPR_data_single_treatment()}} but for more
than one treatment.
}
\details{
If you specify a \code{test_type}, the function will perform a paired t-test or
paired wilcox test and add brackets with significance stars through
\code{ggsignif::geom_signif()}.
}
\examples{
plot_PPR_data_multiple_treatments(
  data = sample_PPR_df,
  include_all_treatments = "yes",
  plot_category = 2,
  baseline_label = "B",
  post_hormone_label = "I",
  included_sexes = "both",
  test_type = "t.test",
  theme_options = sample_theme_options,
  treatment_colour_theme = sample_treatment_names_and_colours
)

}
\seealso{
\code{\link[=plot_PPR_data_single_treatment]{plot_PPR_data_single_treatment()}} to plot changes in PPR for a single treatment. See \code{\link[=make_PPR_data]{make_PPR_data()}} for the function used to create the PPR data.
}
