% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_raw_current_data}
\alias{plot_raw_current_data}
\title{Make raw current plots}
\usage{
plot_raw_current_data(
  data,
  plot_treatment = "Control",
  plot_category = 2,
  current_type = "eEPSC",
  y_variable = "P1",
  pruned = "no",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  hormone_end_time = NULL,
  theme_options,
  male_label = "Male",
  female_label = "Female",
  colour_by_sex = "yes",
  software = "Clampfit",
  x_label = "Time (min)",
  treatment_colour_theme,
  geom_text_family = "",
  geom_point_size = 1,
  geom_point_size_raw_data = 3.5,
  geom_point_alpha = 0.7,
  filename_suffix = "",
  save_plot_png = "no",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{A dataframe containing the raw evoked current data generated from
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. If \code{pruned = "yes"} you must use the \verb{$individual_cells} component of a pruned dataset.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. \code{"Control"}). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{current_type}{A character describing the current type. Allowed values
are \code{"eEPSC"} or \code{"sEPSC"}.}

\item{y_variable}{A character value specifying the variable to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
y_variables are \code{"P1"}, \code{"P1_transformed"}, \code{"mean_P1"} and \code{"PPR"}. \emph{Note}: If you
select \code{"mean_P1"}, you must set the \code{pruned} argument to \code{"yes"}. For
spontaneous currents (\code{current_type = "sEPSC"}), the available y_variables
are \code{"amplitude"} or \code{"frequency"}. NOTE: \code{"frequency"} is only available if \code{pruned = "yes"}.}

\item{pruned}{A character value (\code{"yes"} or \code{"no"}) specifying if the data are
pruned. The plot will then present the data as means with error bars. This is only relevant for y variables like \code{mean_P1} for \code{current_type = "eEPSC"}.}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
\code{"500 nM Insulin"}, \code{"CCK + Leptin"}, and \code{"Insulin"}. If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{hormone_or_HFS_start_time}{A numeric value indicating the time (in
minutes) when a hormone was added or when HFS was applied. This will set
the annotation line start point. This can also be set to \code{NULL} if you want to create your own more complex annotations (for example, for washout, or multiple hormones added at different times). To add your own annotations, use \code{ggplot2::annotate(geom = "segment")} as a starting point.}

\item{hormone_end_time}{A numeric value indicating the time (in minutes) when
a hormone stopped being applied, such as for a washout experiment.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{male_label}{A character value used to describe how males are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for male data in the \code{sex} column, and it must be consistent across data sheets. Defaults to \code{"Male"}.}

\item{female_label}{A character value used to describe how females are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for female data in the \code{sex} column, and it must be consistent across data sheets. This must be consistent in all data sheets. Defaults to \code{"Female"}.}

\item{colour_by_sex}{A character ("yes" or "no") describing if the colour should change based on sex. If "yes", the male data will be coloured according to the \code{colours} column of \code{treatment_colour_theme}, and female data will be coloured according to the \code{very_pale_colours} column of \code{treatment_colour_theme}. If "no", all plots will be coloured using the \code{colours} column.}

\item{software}{A character (\code{"Clampfit"} or \code{"MiniAnalysis"}) describing what software tool was used to analyze the data in \code{new_raw_data_csv}. This is relevant when \code{data_type} is \code{"sEPSC"} because the exported data is different. Defaults to \code{"Clampfit"}.}

\item{x_label}{A character value specifying the x-axis label. Defaults to "Time (min)".}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{geom_text_family}{A character value describing the font family used for the scale bar annotations. Defaults to \code{""} (empty, will use default system font), but can be replaced with a named font. Use a package like \code{extrafont} to load system fonts into R.}

\item{geom_point_size}{A numeric value describing the size of \code{geom_pointrange} for sEPSC pruned amplitude values or \code{geom_point} size for sEPSC raw frequency values. Defaults to \code{1}. Use \code{geom_point_size_raw_data} to specify the point size for all other data types (eEPSC amplitude).}

\item{geom_point_size_raw_data}{A numeric value describing the size of \code{geom_point} for eEPSC data. Defaults to \code{3.5}. If you are plotting sEPSC data, change \code{geom_point_size} instead.}

\item{geom_point_alpha}{A numeric value describing the alpha of \code{geom_point} or \code{geom_pointrange}. Defaults to \code{0.7}.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you want to specify anything about the data (for example, to distinguish between recordings produced in MiniAnalysis
vs. Clampfit).}

\item{save_plot_png}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below \verb{Figures/} in your project directory.}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A list of ggplot objects, where each list element is a scatterplot
of one recording. If \code{save_plot_png == "yes"}, it will also generate a .png
file from each ggplot element in the list. The figures will be exported to
\code{Figures/Evoked-currents/Output-individual-plots} or
\code{Figures/Spontaneous-currents/Output-individual-plots}, depending on the
\code{current_type}. The .png filename will contain the \code{letter}. If the data
are pruned, the filename will also include \verb{_pruned} in the filename. If
the data are normalized (\verb{y_variable == "P1_transformed}), the title will
include \verb{_normalized} in the filename. Example filenames include \code{"A.png"},
\code{"A_normalized"}, and \code{"A_pruned.png"}.
}
\description{
This function enables you to create a scatterplot of raw evoked (or
spontaneous) current amplitude over time in minutes. The plot title will be
pulled from the recording ID (the \code{letter} column in the raw data), and the
subtitle will include the sex and treatment. The plot will also contain a
horizontal line annotation displayed over the time region where a hormone was
applied. If a high-frequency stimulation (HFS) protocol is used, the plot
will display an arrow with the label \code{"HFS"} to indicate the time when HFS was
applied.
}
\examples{

# Plot raw data
plot_raw_current_data(
  data = sample_raw_eEPSC_df,
  plot_treatment = "Control",
  plot_category = 2,
  current_type = "eEPSC",
  y_variable = "P1",
  pruned = "no",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  theme_options = sample_theme_options,
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_plot_png = "no"
)


# Plot pruned data

# Note that this requires the `$individual_cells` element
# of the list generated with `make_pruned_EPSC_data()`.

plot_raw_current_data(
  data = sample_pruned_eEPSC_df$individual_cells,
  plot_treatment = "Control",
  plot_category = 2,
  current_type = "eEPSC",
  y_variable = "mean_P1",
  pruned = "yes",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  theme_options = sample_theme_options,
  treatment_colour_theme = sample_treatment_names_and_colours
)

}
