% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_abf}
\alias{sample_abf}
\title{Sample ABF file}
\format{
An Axon Binary Format (.abf) file with 10000 observations of 5
variables.
\describe{
\item{episode}{A factor representing the sweep, such as "epi1".}
\item{time_sec}{Time in seconds.}
\item{time}{Time in ms*100.}
\item{voltage}{Numeric value representing voltage in mV.}
\item{current}{Numeric value representing current amplitude in pA.}
}
}
\description{
This is an excerpt from a raw recording file exported from Clampfit as an
.abf file. It can be read using the \code{\link[=import_ABF_file]{import_ABF_file()}} function, which will
convert it into a regular dataframe for further manipulation and plotting in
R. It is used to demonstrate the \code{\link[=make_representative_sp_trace]{make_representative_sp_trace()}} function.
}
\examples{
import_ABF_file(import_ext_data("sample_abf.abf"))
}
\keyword{data}
