% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_abf_file}
\alias{sample_abf_file}
\title{Evoked current recording excerpt}
\format{
An Axon Binary Format (.abf) file with 43061 observations of 3
variables.
\describe{
\item{episode}{A factor representing the sweep, such as "epi1".}
\item{current}{Numeric value representing current amplitude in pA.}
\item{time_sec}{Time in seconds.}
}
}
\description{
This is an excerpt from a raw recording of evoked currents exported from
Clampfit as an .abf file. It was then converted into a regular dataframe using the \code{\link[=import_ABF_file]{import_ABF_file()}}
function. To reduce the package installation size, I kept only three columns from the original file. This dataset is used to demonstrate the
\code{\link[=plot_spontaneous_current_trace]{plot_spontaneous_current_trace()}} function.
}
\examples{

plot_spontaneous_current_trace(
  data = sample_abf_file,
  plot_colour = "#6600cc",
  plot_category = 2,
  plot_treatment = "Control",
  state = "Baseline",
  plot_episode = "epi1",
  trace_annotation = "Baseline",
  geom_text_colour = "#6600cc"
)

}
\seealso{
\code{\link[=plot_spontaneous_current_trace]{plot_spontaneous_current_trace()}}
}
\keyword{data}
