% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_AP_count_data}
\alias{sample_AP_count_data}
\title{Example of action potential count data for two time-points: during
the control period (state == "Baseline") or after insulin has been applied
(state == "Insulin"). These were obtained by counting the number of action potentials present within each sweep. The data were then merged with the \code{sample_cell_characteristics} data through \code{add_new_cells()} with \code{data_type = "AP_count"}.}
\format{
An .rda file containing 340 objects of 16 variables.
\describe{
\item{\verb{letter, synapses, ID, sex, animal, treatment, etc.}}{All columns that were produced in
\code{\link[=add_new_cells]{add_new_cells()}}. Please see \link{sample_cell_characteristics} for detailed
column descriptions.}
\item{\code{state}}{A character ("Baseline" or "Insulin") representing
the timepoint that the data point belongs to.}
\item{\code{sweep}}{The sweep.}
\item{\code{no_of_APs}}{The number of action potentials present in the sweep.}
\item{\verb{AP_frequency\}\{The frequency of the action potentials (Hz) in a sweep. This column is generated in }add_new_cells()\verb{.\} \\item\{}current_injection`}{The value of the current injection (in pA)  at a particular sweep.}
}
}
\description{
Example of action potential count data for two time-points: during
the control period (state == "Baseline") or after insulin has been applied
(state == "Insulin"). These were obtained by counting the number of action potentials present within each sweep. The data were then merged with the \code{sample_cell_characteristics} data through \code{add_new_cells()} with \code{data_type = "AP_count"}.
}
\seealso{
\code{\link[=plot_AP_frequencies_multiple_treatments]{plot_AP_frequencies_multiple_treatments()}}
}
\keyword{data}
