% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_eEPSC_t_test_df}
\alias{sample_eEPSC_t_test_df}
\title{Sample eEPSC t-test results}
\format{
An .rda file containing 16 objects of 10 variables
\describe{
\itemize{
\item \code{treatment} A character value specifying the treatment applied (e.g.
"Control", "HNMPA").
\item \code{.y.} A character value describing the parameter compared with the
t-test. In this case, this uses the mean normalized amplitude of the first
evoked current ("mean_P1_transformed") which has been generated using
\code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.
\item \code{group1} A character value describing the first group used in the
paired t-test. This is the baseline interval, "t0to5".
\item \code{group2} A character value describing the second group used in
the paired t-test. Examples include "t5to10" and "t10to15".
\item \code{n1} The number of paired values used in the comparison.
\item \code{statistic} The test statistic.
\item \code{df} The degrees of freedom
\item \code{p_string} The p-value expressed as a character value using
\code{lazyWeave::pvalString()}, which rounds the values and expresses them
according to typical published values (e.g. > 0.99, < 0.001 instead of
exact values)
\item \code{significance_stars} A character value containing asterisks indicating
significance. \verb{p < 0.05 = *}, \verb{p < 0.01 = **}, and \verb{p < 0.001 = ***}.
\item \code{asterisk_time} A numeric value indicating the time value where a
significance star will be added. This is the midpoint of each interval. For
example, the \code{asterisk_time} for an interval from \code{5} to \code{10} minutes will
be \code{7.5} minutes.
}
}
}
\description{
This is an example of the dataframe of t-test results produced using
\code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. The dataframe can be used for further
statistical analyses, and it also has specialized columns to add significance
stars to the plot produced with \code{\link[=make_summary_plot]{make_summary_plot()}}.
}
\keyword{data}
