% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_eEPSC_variance_df}
\alias{sample_eEPSC_variance_df}
\title{Sample evoked currents variance dataframe}
\format{
An .rda file containing 38 objects of 23 variables
\describe{
\itemize{
\item \code{state} A character value describing if a data point belongs to the
baseline interval ("Baseline") or an interval after a hormone or protocol has
been applied ("Post-modification"). These intervals are selected from
\code{baseline_interval} and \code{post_hormone_interval}.
\item \code{mean_cv_inverse_square} The mean inverse coefficient of variation
squared within a specific state.
\item \code{mean_VMR} The mean variance-to-mean ratio within a specific state.
\item \verb{category, letter, sex, treatment, etc.} Other columns which are
imported directly from the summary data without modifications. For a
detailed description of these columns, please see \link{sample_summary_eEPSC_df}.
}
}
}
\description{
This is an example of the dataframe generated from evoked current summmary
data processed using the function \code{\link[=make_variance_data]{make_variance_data()}}. It contains new
columns for the mean variance-to-mean ratio (VMR) and mean inverse
coefficient of variation squared for two time intervals. They are the
baseline period ("t0to5"), and a time interval after a hormone or protocol
has been applied (in this case, "t20to25"). It can be used for further
analyses and also to make the variance comparison plots.
}
\keyword{data}
