% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_pruned_eEPSC_df}
\alias{sample_pruned_eEPSC_df}
\title{Sample pruned eEPSC dataframe}
\format{
An list of 3 dataframes: \code{individual_cells}, \code{for_table} and
\code{all_cells}.
\describe{
\itemize{
\item \code{individual_cells} is used to make pruned plots of individual
recordings using \code{\link[=make_raw_plots]{make_raw_plots()}}. For an explanation of each variable
included in this dataframe, please see the documentation for
\code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.
\item \code{for_table} is used to make the experiment overview table. For an
explanation of each variable included in this dataframe, please see the
documentation for \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.
\item \code{all_cells} is used in \code{\link[=make_summary_plot]{make_summary_plot()}}. For an explanation of
each variable included in this dataframe, please see the documentation for
\code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.
}
}
}
\description{
This is an example of the dataframe of excitatory current amplitudes produced
using \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}. This dataset is a list with three dataframes
containing the pruned data. It can be used for further statistical analyses
or for plotting. \code{\link[=make_raw_plots]{make_raw_plots()}} and \code{\link[=make_summary_plot]{make_summary_plot()}} in particular
depend heavily on this type of dataset.
}
\details{
\itemize{
\item \code{individual_cells} is used to make pruned plots of individual
recordings using \code{\link[=make_raw_plots]{make_raw_plots()}}.
\item \code{for_table} is used to make the experiment overview table
\item \code{all_cells} is used in \code{\link[=make_summary_plot]{make_summary_plot()}}
}
}
\keyword{data}
