% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_pruned_sEPSC_df}
\alias{sample_pruned_sEPSC_df}
\title{Spontaneous current data pruned to a summary point per minute}
\format{
An list of 3 dataframes: \code{individual_cells}, \code{for_table} and
\code{all_cells}.
\describe{
\itemize{
\item \code{individual_cells} is used to make pruned plots of individual
recordings using \code{\link[=plot_raw_current_data]{plot_raw_current_data()}}.
\itemize{
\item \verb{letter, category, sex, treatment, etc.} Information about cell
characteristics, which has automatically been pulled from
cell_characteristics_df in [add_new_cells(). For a definition of these
columns, please see the
\emph{Required basic columns} section in the documentation for
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.
\item \code{interval_pruned} A character value describing the interval that was
used for the pruning function. If the data are pruned per minute, this
will be "t0to1", "t1to2", "t2to3", etc.
\item \code{mean_amplitude} The mean normalized amplitude (\% Baseline sEPSC
amplitude) of all spontaneous currents during a specific interval.
\item \code{mean_raw_amplitude} The mean raw amplitude (pA) of all spontaneous
currents during a specific interval.
\item \code{sd_amplitude} The standard deviation of the normalized spontaneous
current amplitudes.
\item \code{n} The number of data points used.
\item \code{frequency} The frequency of the spontaneous currents (Hz) within
the interval.
\item \code{se} The standard error of the normalized spontaneous current
amplitudes.
\item \code{baseline_range} A logical value required to normalize frequency
relative to the mean baseline frequency. It is set to TRUE when time is
within the baseline period (e.g. \code{time <= 5}) and FALSE at all other times.
Generated automatically in \code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.
\item \code{baseline_mean_frequency} The mean frequency of the spontaneous
currents during the baseline period.
\item \code{frequency_transformed} The normalized spontaneous current frequency
(\% Baseline frequency).
\item \code{time} The upper time value of the interval (e.g. 2 minutes for
"t1to2") which is used on the x-axis of plots such as in
\code{\link[=plot_raw_current_data]{plot_raw_current_data()}}.
}

\item \code{for_table} A dataframe containing two columns: letter and
\code{spont_amplitude_transformed}. The current data is collapsed into a single
row for each letter, with the current data for each letter stored as a list.
This is required to create sparklines of current amplitude over time within
the cell summary table. See \code{\link[=make_interactive_summary_table]{make_interactive_summary_table()}}.

\item \code{all_cells} A dataframe consisting of all data within a single
treatment grouped and summarized per minute (or some other variable if you
change \code{interval_length} to be something other than \code{1}). Columns like
\code{category} and \code{sex} are retained from the raw data. New columns for
spontaneous current data (\code{current_type == "sEPSC"}) are:
\itemize{
\item \code{interval_pruned} A character value describing the interval that was
used for the pruning function. If the data are pruned per minute, this
will be "t0to1", "t1to2", "t2to3", etc.
\item \code{mean_all_amplitude} The mean normalized amplitude (\% Baseline
sEPSC amplitude) of the spontaneous current amplitudes during a specific
interval across all cells.
\item \code{mean_all_raw_amplitude} The mean raw amplitude (in pA) of the
spontaneous current amplitudes during a specific interval across all
cells.
\item \code{sd_all_amplitude} The standard deviation of the normalized
spontaneous current amplitudes across all cells.
\item \code{n} The number of data points used.
\item \code{se_all_amplitude} The standard error of the normalized spontaneous
currents across all cells.
\item \code{sd_all_raw_amplitude} The standard deviation of the raw
spontaneous current amplitudes across all cells.
\item \code{se_raw_amplitude} The standard error of the raw spontaneous
currents across all cells.
\item \code{mean_all_frequency} The mean normalized frequency (\% Baseline
frequency) of all spontaneous current amplitudes across all cells during
the interval.
\item \code{sd_all_frequency} The standard deviation of the normalized
frequency of all spontaneous current amplitudes across all cells during
the interval.
\item \code{se_frequency} The standard error of the normalized
frequency of all spontaneous current amplitudes across all cells during
the interval.
\item \code{mean_all_raw_frequency} The mean raw frequency (Hz) of all
spontaneous current amplitudes across all cells during the interval.
\item \code{sd_all_raw_frequency} The standard deviation of the raw
frequency of all spontaneous current amplitudes across all cells during
the interval.
\item \code{se_raw_frequency} The standard error of the raw
frequency of all spontaneous current amplitudes across all cells during
the interval.
}
}
}
}
\description{
This is an example of the dataframe of spontaneous currents
produced using \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}. This dataset is a list with three
dataframes containing the pruned data. It can be used for further statistical
analyses or for plotting. \code{\link[=plot_raw_current_data]{plot_raw_current_data()}} and
\code{\link[=plot_summary_current_data]{plot_summary_current_data()}} in particular depend heavily on this type of
dataset.
}
\details{
\itemize{
\item \code{individual_cells} is used to make pruned plots of individual
recordings using \code{\link[=plot_raw_current_data]{plot_raw_current_data()}}.
\item \code{for_table} is used to make the experiment overview table.
\item \code{all_cells} is used in \code{\link[=plot_summary_current_data]{plot_summary_current_data()}}.
}
}
\keyword{data}
