% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_raw_eEPSC_df}
\alias{sample_raw_eEPSC_df}
\title{Sample raw eEPSC data}
\format{
An .rda file containing 5680 objects of 20 variables.
\describe{
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{synapses} A character value (e.g. "Glutamate" or "GABA").
\item \code{sex} A character value (e.g. "Male" or "Female").
\item \code{treatment} A character value (e.g. "Control", "HNMPA").
\item \code{time} A numeric value that represents time in minutes. This column is
autogenerated in \code{\link[=add_new_cells]{add_new_cells()}}.
\item \code{ID} A character value for the recording filename.
\item \code{P1} A numeric value representing the amplitude of the first evoked
current in pA.
\item \code{P2} A numeric value representing the amplitude of the second evoked
current in pA.
\item \code{X} A numeric value representing the x-value of the cell's location in
µm. Leave this blank if you don't have this data.
\item \code{Y} A numeric value representing the y-value of the cell's location in
µm. Leave this blank if you don't have this data.
\item \code{age} A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages).
\item \code{animal} A numeric value representing the animal's ID or number.
\item \code{category} A numeric value representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups.
\item \code{cell} A character or numeric value representing the cell. For
example, use \verb{3.1.1} for animal #3, slice #1, cell #1.
\item \code{PPR} (for evoked currents only) A numeric value that represents the
paired pulse ratio (PPR) of the evoked currents, generated using
\code{dplyr::mutate(PPR = P2/P1)}.
\item \code{interval} A character value indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \code{baseline_range} A logical value required for the baseline
transformation. It is set to TRUE when time is within the baseline period
(e.g. Time <= 5) and FALSE at all other times.
\item \code{baseline_mean} A numeric value representing the mean evoked current
amplitude during the baseline period. There is a different baseline_mean for
each letter.
\item \code{P1_transformed} A numeric value representing the first evoked current
amplitude (pA) normalized relative to the mean amplitude during the
recording's baseline.
\item \code{P2_transformed} A numeric value representing the second evoked
current amplitude (pA) normalized relative to the mean amplitude during the
recording's baseline.
}
}
}
\description{
This is an example of the raw eEPSC data produced using
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. It is useful for demonstrating functions that
build off of this dataset, such as \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}} and plotting
functions.
}
\seealso{
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} and \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}
}
\keyword{data}
