% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_summary_eEPSC_df}
\alias{sample_summary_eEPSC_df}
\title{Evoked current data summarized into 5-minute intervals}
\format{
An .rda file containing 95 objects of 20 variables.
\describe{
\itemize{
\item \code{category} A factor representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups.
\item \code{letter} A factor that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{sex} A factor with two levels (e.g. "Male" or "Female").
\item \code{treatment} A factor based on the treatment applied (e.g. "Control",
"HNMPA"). This represents the antagonists or agonists applied, or any
protocol applied to the animals (e.g. "Fasting").
\item \code{interval} A factor indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \code{mean_P1_transformed} A numeric value representing the mean current
amplitude (in pA normalized to the baseline amplitude) of the first evoked
current. The values used to produce mean_P1_transformed come from all data
points within each interval (the length of the interval was specified in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}).
\item \code{mean_P1_raw} The same values as those in \code{mean_P1_transformed} except
these values contain raw values for current amplitude (e.g. the data were
not baseline transformed).
\item \code{n} The number of data points used to produce each value in
\code{mean_P1_transformed} and \code{mean_P1_raw}
\item \code{sd} The standard deviation of the mean current values.
\item \code{cv} The coefficient of variation of the mean current values.
\item \code{se} The standard error of the mean current values.
\item \code{cv_inverse_square} The inverse coefficient of variation squared of
the mean current values. Calculated by using (\code{1/cv^2}), where \code{cv} is the
coefficient of variation. Useful for variance analysis. \item \code{variance} The
variance of the current values, calculated using \code{stats::var()}.
\item \code{VMR} The variance-to-mean ratio, calculated using
\code{variance/mean_P1_transformed}. Useful for variance analysis.
\item \code{age} A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages).
\item \code{animal} A numeric value representing the animal's ID or number.
\item \code{X} A numeric value representing the x-value of the cell's location in
µm. Leave this blank if you don't have this data.
\item \code{Y} A numeric value representing the y-value of the cell's location in
µm. Leave this blank if you don't have this data.
\item \code{time} A numeric value that represents time in minutes. This column is
autogenerated in add_new_cells()].
\item \code{synapses} A character value (e.g. "Glutamate" or "GABA").
}
}
}
\description{
This is an example of the summary eEPSC data produced using
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. It is useful for demonstrating functions that
build off of this dataset, such as \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}} and plotting
functions such as \code{\link[=plot_summary_current_data]{plot_summary_current_data()}}.
}
\seealso{
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} and \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}
}
\keyword{data}
