% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_summary_sEPSC_df}
\alias{sample_summary_sEPSC_df}
\title{Spontaneous current data summarized into 5-minute intervals}
\format{
An .rda file containing 70 objects of 16 variables.
\describe{
\itemize{
\item \code{category} A factor representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups.
\item \code{letter} A factor that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{sex} A factor with two levels (e.g. "Male" or "Female").
\item \code{treatment} A factor based on the treatment applied (e.g. "Control",
"HNMPA"). This represents the antagonists or agonists applied, or any
protocol applied to the animals (e.g. "Fasting").
\item \code{interval} A factor indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \code{mean_transformed_amplitude} The average normalized spontaneous
current amplitude (\% Baseline sEPSC amplitude).
\item \code{mean_raw_amplitude} The average raw spontaneous current amplitude
(pA).
\item \code{n} The number of datapoints used to create the average.
\item \code{sd_transformed_amplitude} The standard deviation of the normalized
spontaneous current data (\code{mean_transformed_amplitude}).
\item \code{se_transformed_amplitude} The standard error of
\code{mean_transformed_amplitude}.
\item \code{mean_transformed_frequency} The average normalized frequency (\%
Baseline frequency).
\item \code{sd_transformed_frequency} The standard deviation of \code{mean_transformed_frequency}.
\item \code{se_frequency} The standard error of \code{mean_transformed_frequency}.
\item \code{mean_raw_frequency} The average raw frequency (Hz).
\item \code{time} A numeric value that represents time in minutes. This column is
autogenerated in add_new_cells()].
\item \code{synapses} A character value (e.g. "Glutamate" or "GABA").
}
}
}
\description{
This is an example of the summary eEPSC data produced using
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}. It is useful for demonstrating functions that
build off of this dataset, such as \code{\link[=plot_summary_current_data]{plot_summary_current_data()}}.
}
\seealso{
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} and \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}
}
\keyword{data}
