% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_theme_options}
\alias{sample_theme_options}
\title{A dataframe of theme options for things like colours and line widths}
\format{
An .rda file containing 23 objects of 2 variables

\describe{
\item{\code{gray_shading_colour}}{Hex code for the colour used to fill the violin
plots.}
\item{\code{line_col}}{Hex code for the colour used for the line indicating
periods when a hormone was applied.}
\item{\code{rectangle_shading_colour}}{Hex code for the colour used to shade the
rectangles.}
\item{\code{male_shape}}{Numeric value representing the point shape used for data
values where \code{Sex == Male}.}
\item{\code{female_shape}}{Numeric value representing the point shape used for
data values where \code{Sex == Female}.}
\item{\code{mean_point_colour}}{Hex code for the colour used to fill the mean data point in plots like the PPR comparison plot.}
\item{\code{connecting_line_width}}{The width of the connecting line in action potential comparison plots.}
\item{\code{connecting_line_colour}}{The colour of the connecting line in action potential comparison plots.}
\item{\code{connecting_line_width_PPR}}{The width of the connecting line in paired-pulse ratio comparison plots.}
\item{\code{mean_point_size}}{The size of the mean data point in plots like the PPR comparison plot.}
\item{\code{geom_signif_text_size}}{The size of the asterisk text in plots like the spontaneous current comparison plots.}
\item{\code{AP_trace_size}}{DO NOT USE; this will be moved to a function argument in future relases.}
\item{\code{y_axis_limit}}{The maximum y-axis value to be used in plots.}
\item{\code{line_col}}{A character value with the same treatment names as in
\code{treatment} except you would replace underscores with spaces and add line
breaks if needed. This is to create more attractive, human-readable axis
labels and table values.}
}
}
\description{
This is an example of the dataframe used to specify the colour themes and
plot options (like line thickness).
}
\keyword{data}
