% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_treatment_names_and_colours}
\alias{sample_treatment_names_and_colours}
\title{A dataframe of treatments and their assigned colours for consistency across plots}
\format{
An .rda file containing 4 objects of 5 variables
\describe{
\item{\code{category}}{A numeric value representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as
subgroups. For example, "1" may refer to an experiment where you applied high-frequency stimulation (HFS) to a cell, while "2" is an experiment where you added a hormone like leptin. "3" may be an experiment where you applied HFS in the continuous presence of leptin.}
\item{\code{treatment}}{A character value specifying the treatment applied (e.g.
"Control", "HNMPA") such as antagonists or agonists, or any protocol applied
to the animals (e.g. "Fasting").}
\item{\code{display_names}}{A character value with the same treatment names as in
\code{treatment} except you would replace underscores with spaces and add line
breaks if needed. This is to create more attractive, human-readable axis
labels and table values.}
\item{\code{colours}}{A character value with hex colour values.}
\item{\code{very_pale_colours}}{A character value with hex colour values. These
are all lighter than the colours in \code{colours} and they are used to create
contrast in summary plots with both sexes.}
}
}
\description{
This is an example of the dataframe used to specify a colour theme for each
treatment. It is used to create a consistent colour scheme across all plots
produced with this package.
}
\section{Defining your own treatment names and colours}{


Use the data.frame() function to define an object, then refer to this named object in all \code{treatment_colour_theme} arguments. E.g.

\verb{my_theme_colours <- data.frame(}
\verb{category = c(2, 2, 2, 2),}
\verb{treatment = c("Control", "HNMPA", "PPP", "PPP_and_HNMPA"),}
\verb{display_names = c("Control", "HNMPA", "PPP", "PPP\\n&\\nHNMPA"),}
\verb{colours = c("#f07e05", "#f50599", "#70008c", "#DCE319"),}
\code{very_pale_colours = c("#fabb78", "#fa98d5", "#ce90de", "yellow")}
\verb{)}
}

\keyword{data}
