% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{add_new_cells}
\alias{add_new_cells}
\title{Add new data}
\usage{
add_new_cells(
  new_raw_data_csv,
  cell_characteristics_csv,
  old_raw_data_csv,
  current_type,
  new_file_name
)
}
\arguments{
\item{new_raw_data_csv}{A filepath to a csv containing the new raw data. If
the data are evoked current data (\code{current_type == "eEPSC"}) then this must
contain 4 columns: \code{Letter}, \code{ID}, \code{P1} and \code{P2}. If the data are
spontaneous current data, the columns must be \code{Letter}, \code{ID},
\code{Recording_num}, \code{Trace}, \code{Peak_amplitude} and \code{Time_of_peak}. Please see
the section below on required columns for more details.}

\item{cell_characteristics_csv}{A filepath to a csv containing information
about the cells. Please see \code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}} for a
description of what columns should be included.}

\item{old_raw_data_csv}{A filepath to a csv containing the current raw data.
Since this function appends the new data to the old data, this must be of
the same current_type as the new data (e.g. the columns must be the same).}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{new_file_name}{A filename for the csv containing the new data appended
to the old data. Must be a character representing a filepath to a csv file.
Examples include "Data/20241118-Raw-eEPSC-data.csv".}
}
\value{
A dataframe consisting of the old raw data with information from the new
cells appended to it.
}
\description{
Add new data
}
\examples{
\dontrun{
add_new_cells(
  new_raw_data_csv = import_ext_data("sample_new_eEPSC_data.csv"),
  cell_characteristics_csv = import_ext_data("sample_cell_characteristics.csv"),
  old_raw_data_csv = import_ext_data("sample_eEPSC_data.csv"),
  current_type = "eEPSC",
  new_file_name = "20241118-Raw-eEPSC-Data.csv"
)
}

}
