% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{add_new_cells}
\alias{add_new_cells}
\title{Add new data}
\usage{
add_new_cells(
  new_raw_data_csv,
  cell_characteristics_csv,
  old_raw_data_csv,
  data_type,
  write_new_csv = "yes",
  new_file_name,
  decimal_places = 2,
  injection_start_time = 265.4
)
}
\arguments{
\item{new_raw_data_csv}{A filepath to a csv containing the new raw data. If
the data are evoked current data (\code{current_type == "eEPSC"}) then this must
contain 4 columns: \code{letter}, \code{ID}, \code{P1} and \code{P2}. If the data are
spontaneous current data, the columns must be \code{letter}, \code{ID},
\code{recording_num}, \code{trace}, \code{peak_amplitude} and \code{time_of_peak}. Please see
the section below on required columns for more details.}

\item{cell_characteristics_csv}{A filepath to a csv containing information
about the cells. Please see \code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}} for a
description of what columns should be included. Don't forget to update this
to include the cell characteristics for the new letters in
\code{new_raw_data_csv}!}

\item{old_raw_data_csv}{A filepath to a csv containing the current raw data.
Since this function appends the new data to the old data, this must be of
the same current_type as the new data (e.g. the columns must be the same).
If this is the first time you are running this function, start with a blank
.csv file containing just the column titles in the first row.}

\item{data_type}{A character ("eEPSC", "sEPSC" or "AP") describing the data type that is being imported.}

\item{write_new_csv}{A character ("yes" or "no") describing if the new data
should be written to a csv file. Defaults to "yes". Please specify
a filename for the new csv file in \code{new_file_name}.}

\item{new_file_name}{A filename for the csv containing the new data appended
to the old data. Must be a character representing a filepath to a csv file.
Examples include "Data/20241118-Raw-eEPSC-data.csv".}

\item{decimal_places}{A numeric value indicating the number of decimal places the data should be rounded to. Used to reduce file size and prevent an incorrect representation of the number of significant digits.}

\item{injection_start_time}{For action potential data only: A numeric value describing the start time (in ms) when current injection was applied. Used to calculate the latency to fire.}
}
\value{
A dataframe consisting of the old raw data with information from the new
cells appended to it. If \code{data_type == "AP"} two new columns will also be
added based on calculations from the existing columns. These are
\code{latency_to_fire} (which is \code{time_to_peak} - \code{injection_start_time}) and
\code{antipeak_time_relative_to_threshold} (which is \code{time_of_antipeak} -
\code{time_of_threshold}).
}
\description{
This function enables you to append new raw recording data onto an existing
datasheet. It makes it easy and convenient to merge the cell parameters (age,
sex, etc.) with new data and add it to your current raw data. This function
also formats the dataset so it is immediately ready for use in functions like
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.
}
\section{Required Columns}{


If the data are evoked currents (\code{data_type == "eEPSC"}), the data must
contain the following four columns:

\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{P1} A numeric value representing the
amplitude of the first evoked current in pA.
\item \code{P2} A numeric value representing the
amplitude of the second evoked current in pA.
}

If the data are spontaneous currents (\code{data_type == "sEPSC"}), the data
must contain the following columns:
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{recording_num} A numeric value representing the recording number.
This was incorporated before we switched to concatenating all recordings
into one, but it needs to remain here to prevent breaking previous projects.
It should be set to 1.
\item \code{trace} A numeric value representing the trace (automatically
generated in Clampfit) where the current occurred.
\item \code{time_of_peak} A numeric value representing the time of the peak in
milliseconds relative to trace number. This is automatically calculated in
Clampfit.
\item \code{time} A numeric value representing the absolute time when the current
happened, relative to the start of the recording. This is autogenerated. See
\code{\link[=add_new_cells]{add_new_cells()}} for a description of how the true time value (\code{time}) is
calculated from the \code{recording_num} and \code{trace.}
\item \code{amplitude} A numeric value representing the amplitude of the evoked
current in pA.
}

If the data are action potential parameters (\code{data_type == "AP"}), the data
must contain the following columns:
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.
\item \code{ID} A character value for the recording filename.
\item \code{state} A character ("Baseline" or "Insulin") representing
the timepoint that the data point belongs to.
\item \code{time_of_threshold} The time (in ms) when the membrane potential
reaches the threshold value.
\item \code{threshold} The threshold (in mV). Determined using the first
derivative method, where the threshold is the membrane potential which
results in a derivative of 10 mV/ms or greater (Farries et al., 2010).
\item \code{t_11} The value of the first derivative (action potential velocity in
mV/ms) at threshold.
\item \code{first_sweep_with_APs} The sweep number of the first sweep (going from
lowest to higher current injection values) that resulted in an action
potential.
\item \code{trace_start} An automated value in Clampfit that is not used in the
analysis.
\item \code{peak_amplitude} The peak amplitude (in pA) of the action
potential relative to threshold.
\item \code{time_to_peak} The time to peak
amplitude (in ms) relative to the time of threshold.
\item \code{antipeak_amplitude} The after-hyperpolarization amplitude (in pA) relative
to threshold.
\item \code{time_of_antipeak} The time of the after-hyperpolarization (in ms).
\item \code{half_width} The half-width, which is the width of the action potential and half of the peak amplitude.
}
}

\examples{
\dontrun{
add_new_cells(
  new_raw_data_csv = import_ext_data("sample_new_eEPSC_data.csv"),
  cell_characteristics_csv = import_ext_data("sample_cell_characteristics.csv"),
  old_raw_data_csv = import_ext_data("sample_eEPSC_data.csv"),
  data_type = "eEPSC",
  write_new_csv = "no",
  new_file_name = "20241118-Raw-eEPSC-Data.csv",
  decimal_places = 2
)
}

}
\seealso{
\code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}} for a list of required columns in the \code{cell_characteristics_csv}.

\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} for the next step in the analysis process.
}
