% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read-and-plot-raw-ABF-files.R
\name{import_ABF_file}
\alias{import_ABF_file}
\title{Import raw .abf files as a dataframe}
\usage{
import_ABF_file(file_name)
}
\arguments{
\item{file_name}{Filepath to an .abf file (e.g. "Data/23711004.abf")}
}
\value{
A dataframe
}
\description{
\code{import_ABF_file()} is a wrapper around \code{abftools::abf2_load()} and
\code{abftools::MeltAbf()}. It converts the array from \code{abf2_load()} into a
dataframe, and it also converts time to minutes
}
\examples{
import_ABF_file(load_ext_data("sample_abf.abf"))
}
