% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{import_ABF_file}
\alias{import_ABF_file}
\title{Import raw \code{.abf} files as a dataframe}
\usage{
import_ABF_file(file_name, recording_mode)
}
\arguments{
\item{file_name}{Filepath to an \code{.abf} file (e.g. "Data/23711004.abf")}

\item{recording_mode}{The mode used for the recording. If in \code{recording_mode = "voltage_clamp"} (e.g. clamping cell at -70 mV and measuring current amplitude) the primary channel (\code{chan1}) is set to "current" and the secondary channel (\code{chan2}) is "voltage". If \code{recording_mode = "current_clamp"}, these values are reversed, where the primary channel is "voltage" and the secondary channel is "current".}
}
\value{
A dataframe with 5 columns:

\itemize{
\item \code{time} Time value from Clampfit which is in milliseconds x 10. For example, 5 seconds = 50000 in this column.
\item \code{episode} Character value (e.g. "epi1", "epi2") which corresponds
directly to "sweep" in Clampfit.
\item \code{current} Current in pA.
\item \code{voltage} Voltage in mV.
\item \code{time_sec} Time in seconds.
}
}
\description{
\code{import_ABF_file()} is a wrapper around \code{abftools::abf2_load()} and
\code{abftools::MeltAbf()}. It converts the array from \code{abf2_load()} into a
dataframe, and it also converts time to minutes.
}
\details{
The file progresses from \code{.abf} to an array and then to a dataframe that can be
easily manipulated in R. Be sure to assign a value to the dataframe so you can use it in later functions.
}
\examples{
import_ABF_file(import_ext_data("sample_abf.abf"), recording_mode = "voltage_clamp")
}
