% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import-data.R
\name{import_cell_characteristics_df}
\alias{import_cell_characteristics_df}
\title{Import cell characteristics}
\usage{
import_cell_characteristics_df(filename)
}
\arguments{
\item{filename}{A filepath to a .csv file, written as a character. The
function uses \code{here::here()} to locate the filepath. See the details below
for information on required columns.}
}
\value{
A dataframe
}
\description{
\code{import_cell_characteristics_df()} is a wrapper around \code{read.csv()} to import
a .csv file with information about a cell (animal, age, sex, synapses, X- and
Y-coordinates, etc.). It replaces \code{NA} values in the \code{R_a} column with \code{0} to
remove errors caused by missing data. The resulting dataframe can be merged
with raw data into a summary table and used in downstream statistical
analyses.
}
\section{Required columns}{


These columns are required in the raw .csv file:
\itemize{
\item \code{letter} A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics. Example: "A"
\item \code{cell} A character or numeric value representing the cell. For
example, use \verb{3.1.1} for animal #3, slice #1, cell #1.
\item \code{sex} A character value such as "Male" or "Female".
\item \code{X} A numeric value representing the x-value of the cell's location in
µm.
\item \code{Y} A numeric value representing the y-value of the cell's location in
µm.
\item \code{age} A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages). Do not use characters (e.g. avoid
"P31" and use 31 instead).
\item \code{animal} A numeric value representing the animal's ID or number.
\item \code{synapses} A character value such as "Glutamate" or "GABA".
\item \code{treatment} A character value such as "Control" or "HNMPA".
\item \code{category} A numeric value representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups.
\item \code{R_a} A list of values for the access resistance, which would have been
monitored at several timepoints throughout the recording. See the section
\code{R_a} formatting below. }
}

\section{\code{R_a} formatting}{


\code{R_a} is a mandatory column with information about the cell's access
resistance. Each element of this column must be a sequence of numbers,
separated by a comma and a single space. Although this will be read in as a
character, do not add quotation marks around the values in this column. For
example, \verb{1.5, 1.5, 1.6, 1.7, 1.7, 1.8} is an acceptable \code{R_a} value for a
single cell.

\code{import_cell_characteristics_df()} will convert the character value into a
list of numeric values (using \code{stringr::str_split()}). It will also convert
blanks and \code{NA} values to 0. This allows access to be visualized as a
sparkline in the \code{R_a} column of the interactive summary table made with
make_interactive_summary_table().
}

\examples{
import_cell_characteristics_df(import_ext_data("sample_cell_characteristics.csv"))

}
\seealso{
make_interactive_summary_table() to generate an interactive table
with cell characteristics and raw data as sparklines.
}
