% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{make_interactive_summary_table}
\alias{make_interactive_summary_table}
\title{Make interactive overview table of all recordings}
\usage{
make_interactive_summary_table(
  cell_characteristics_dataframe,
  pruned_eEPSC_dataframe,
  pruned_sEPSC_dataframe,
  treatment_colour_theme,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  include_all_categories = "yes",
  list_of_categories = NULL,
  save_output_as_RDS = "no"
)
}
\arguments{
\item{cell_characteristics_dataframe}{A dataframe containing the cell characteristics, generated from \code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}}.}

\item{pruned_eEPSC_dataframe}{A dataframe containing pruned evoked current
data, generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}, where \code{current_type == "eEPSC"}.}

\item{pruned_sEPSC_dataframe}{A dataframe containing pruned spontaneous
current data, generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}, where \code{current_type == "sEPSC"}.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{include_all_treatments}{A character ("yes" or "no") specifying if the
plot will include data from all treatments. If "no", you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to NULL, since
include_all_treatments is "yes" by default.}

\item{include_all_categories}{A character ("yes" or "no") specifying if the
plot will include data from all categories. If "no", you must specify a
list of categories in \code{list_of_categories}.}

\item{list_of_categories}{A list of character values describing the
categories that will be in the plot. Defaults to NULL, since
include_all_categories is "yes" by default.}

\item{save_output_as_RDS}{A character ("yes" or "no") describing if the
resulting object should be saved as an RDS file in the raw data folder.
Note: This is not the interactive table, but it is the raw dataframe that
is later inserted into \code{reactable::reactable()}. This is useful if you want
to build your own table using a different package, or you want to generate
a customized reactable table yourself.}
}
\value{
A reactable HTML widget that can be viewed in RStudio or exported in
RMarkdown HTML documents. If \code{save_output_as_RDS == "yes"}, the raw
dataframe used to create the reactable is also exported as an .rds file
into \verb{Data/Output-Data-from-R/}
}
\description{
This function pulls information from multiple dataframes to display
everything about a cell (cell characteristics, evoked current data, and
spontaneous current data) in an interactive table. The table is made with
\code{reactable::reactable()}, so it can be filtered, sorted, and rearranged.
}
\details{
The table contains sparklines of the evoked current and spontaneous current
amplitudes over time, which allows you to have a quick overview of the
responses of a cell, and d

The sparklines are colour-coded by treatment, allowing you to quickly
identify trends in response to a hormone/protocol for all cells belonging to
a particular treatment.
}
\examples{

make_interactive_summary_table(
  cell_characteristics_dataframe = sample_cell_characteristics,
  pruned_eEPSC_dataframe = sample_pruned_eEPSC_df,
  pruned_sEPSC_dataframe = sample_pruned_sEPSC_df,
  treatment_colour_theme = sample_treatment_names_and_colours,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  include_all_categories = "yes",
  list_of_categories = NULL,
  save_output_as_RDS = "no"
)

}
