% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{make_summary_EPSC_data}
\alias{make_summary_EPSC_data}
\title{Summarize current data per 5-min for statistical tests}
\usage{
make_summary_EPSC_data(
  data = patchclampplotteR::sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no"
)
}
\arguments{
\item{data}{A \code{data.frame} object. If the data are evoked currents
(\code{current_type == "eEPSC"}), this should be the raw evoked current data
generated using \code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. If the data are spontaneous
currents (\code{current_type == "sEPSC"}), this should be the pruned data
\verb{$individual_cells} dataset generated using \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{save_output_as_RDS}{A character ("yes" or "no") describing if the
resulting object should be saved as an RDS file in the raw data folder.}
}
\value{
A dataframe with summary data such as the mean current amplitude,
coefficient of variation, standard deviation, standard error, variance,
variance-to-mean ratio, and inverse coefficient of variation squared for
each interval.

New columns for evoked current data (\code{current_type == "eEPSC"}) include:
\itemize{
\item \code{mean_P1_transformed} The amplitude of the first evoked current
amplitude (\% Baseline eEPSC amplitude) normalized to the mean baseline
amplitude and averaged over the interval.
\item \code{mean_P1_raw} The amplitude of the first evoked current amplitude (pA)
averaged over the interval.
\item \code{n} The number of datapoints used to create the averaged values.
Corresponds to the number of sweeps per interval.
\item \code{sd} The standard deviation of the normalized evoked current data
(\code{P1_transformed}).
\item \code{cv} The coefficient of variation of \code{P1_transformed}.
\item \code{se} The standard error of \code{P1_transformed}.
\item \code{cv_inverse_square} The inverse of the squared coefficient of
variation of \code{P1_transformed}.
\item \code{variance} The variance of \code{P1_transformed}.
\item \code{VMR} The variance-to-mean ratio (VMR) of \code{P1_transformed}.
\item \code{interval} A character value indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \verb{letter, synapses, sex, treatment, etc.} Unmodified columns from the
original dataset describing the cell's properties.
}

New columns for spontaneous current data (\code{current_type == "sEPSC"}) include:

\itemize{
\item \code{mean_transformed_amplitude} The average normalized spontaneous
current amplitude (\% Baseline sEPSC amplitude).
\item \code{mean_raw_amplitude} The average raw spontaneous current amplitude
(pA).
\item \code{n} The number of datapoints used to create the average.
\item \code{sd_transformed_amplitude} The standard deviation of the normalized
spontaneous current data (\code{mean_transformed_amplitude}).
\item \code{se_transformed_amplitude} The standard error of
\code{mean_transformed_amplitude}.
\item \code{mean_transformed_frequency} The average normalized frequency (\%
Baseline frequency).
\item \code{sd_transformed_frequency} The standard deviation of \code{mean_transformed_frequency}.
\item \code{se_frequency} The standard error of \code{mean_transformed_frequency}.
\item \code{mean_raw_frequency} The average raw frequency (Hz).
\item \verb{letter, synapses, sex, treatment, etc.} Unmodified columns from the
original dataset describing the cell's properties.
}
}
\description{
\code{make_summary_EPSC_data()} allows you to divide data from a long recording
(e.g. 30 minutes) into evenly-spaced intervals (e.g. 5 minutes). It will
generate summary data like the mean current amplitude for each interval. This
can be useful for inserting into statistical models to compare effect sizes
across broad stretches of time. The interval length would have been
previously specified in \code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} using the
\code{interval_length} argument.
}
\examples{

make_summary_EPSC_data(
  data = sample_raw_eEPSC_df,
  current_type = "eEPSC",
  save_output_as_RDS = "no"
)

make_summary_EPSC_data(
  data = sample_pruned_sEPSC_df$individual_cells,
  current_type = "sEPSC",
  save_output_as_RDS = "no"
)


}
