% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{make_variance_data}
\alias{make_variance_data}
\title{Make dataframe with variance measures}
\usage{
make_variance_data(
  data,
  include_all_categories = "yes",
  list_of_categories = NULL,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  treatment_colour_theme,
  save_output_as_RDS = "no"
)
}
\arguments{
\item{data}{A dataframe containing the summary data generated from
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}. If \code{current_type} is "eEPSC", this must be the \verb{$summary_data} element of the list produced by \code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}.}

\item{include_all_categories}{A character (\code{"yes"} or \code{"no"}) specifying if the
plot will include data from all categories. If \code{"no"}, you must specify a
list of categories in \code{list_of_categories}.}

\item{list_of_categories}{A list of character values describing the
categories that will be in the plot. Defaults to \code{NULL}, since
\code{include_all_categories} is \code{"yes"} by default.}

\item{include_all_treatments}{A character (\code{"yes"} or \code{"no"}) specifying if the
plot will include data from all treatments. If \code{"no"}, you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to \code{NULL}, since
include_all_treatments is \code{"yes"} by default.}

\item{baseline_interval}{A character value indicating the name of the
interval used as the baseline. Defaults to \code{"t0to5"}, but can be changed.
Make sure that this matches the baseline interval that you set in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{post_hormone_interval}{A character value indicating the name of the
interval used as "after" timepoint for comparison. Defaults to \code{"t20to25"},
but can be changed. Make sure that this matches an interval present in
\code{data}}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_output_as_RDS}{A character (\code{"yes"} or \code{"no"}) describing if the
resulting object should be saved as an RDS file in the folder \code{"Data/Output-Data-from-R"}. The function will automatically create this folder if it doesn't already exist.
Note: This is not the interactive table, but it is the raw dataframe that
is later inserted into \code{reactable::reactable()}. This is useful if you want
to build your own table using a different package, or you want to generate
a customized reactable table yourself.}
}
\value{
A dataframe containing all of the columns within the summary data
(see \link{sample_summary_eEPSC_df} for a detailed description of these columns)
plus three additional columns:
\itemize{
\item \code{state} A character value describing if a data point belongs to the
baseline interval (\code{"Baseline"}) or an interval after a hormone or protocol has
been applied (\code{"Post-modification"}). These intervals are selected from
\code{baseline_interval} and \code{post_hormone_interval}.
\item \code{mean_cv_inverse_square} The mean inverse coefficient of variation
squared within a specific state and sex.
\item \code{mean_VMR} The mean variance-to-mean ratio within a specific state and sex.
}
}
\description{
\code{make_variance_data} creates a dataframe containing variance measures at two
time points. They are the baseline period and a user-specified interval after
a hormone or protocol has been applied. The variance measures are the inverse
coefficient of variation squared and the variance-to-mean ratio (VMR). A
\code{"before vs. after"} comparison of these two variance measures is useful to
determine which mechanism is involved in modifying synaptic plasticity. For
more information, please see \href{https://doi.org/10.1016/j.jneumeth.2019.108526}{Huijstee & Kessels (2020)}.
}
\examples{

make_variance_data(
  data = sample_summary_eEPSC_df$summary_data,
  include_all_categories = "no",
  list_of_categories = c(2),
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_output_as_RDS = "no"
)

}
\seealso{
\code{\link[=plot_variance_comparison_data]{plot_variance_comparison_data()}} to plot this data.
}
