% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{patchclampplotteR_facet_theme}
\alias{patchclampplotteR_facet_theme}
\title{The patchclampplotteR theme for facet plots}
\usage{
patchclampplotteR_facet_theme()
}
\value{
A ggplot theme
}
\description{
This is a modified version of \code{\link[=patchclampplotteR_theme]{patchclampplotteR_theme()}} that is optimized for plots produced with \code{\link[=make_facet_plot]{make_facet_plot()}}. It features large, easy-to-read axis labels and titles for individual facets, and margins that allow for easier readability.
}
\seealso{
\code{\link[=patchclampplotteR_theme]{patchclampplotteR_theme()}} for the original theme and \code{\link[=make_facet_plot]{make_facet_plot()}}.
}
