% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-data.R
\name{perform_t_tests_for_summary_plot}
\alias{perform_t_tests_for_summary_plot}
\title{Perform t-tests for EPSC summary plots}
\usage{
perform_t_tests_for_summary_plot(
  data,
  test_category,
  include_all_treatments = "yes",
  list_of_treatments = NULL,
  treatment_colour_theme,
  baseline_interval,
  interval_length,
  parameter,
  current_type,
  save_output_as_RDS = "no"
)
}
\arguments{
\item{data}{A dataframe containing the summary data generated from
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}.}

\item{test_category}{A numeric value describing the experimental category. In
the sample dataset for this package, 2 represents experiments where insulin
was applied continuously after a 5-minute baseline period.}

\item{include_all_treatments}{A character ("yes" or "no") specifying if the
plot will include data from all treatments. If "no", you must specify a
list of treatments in \code{list_of_treatments}.}

\item{list_of_treatments}{A list of character values describing the
treatments that will be in the plot. Defaults to NULL, since
include_all_treatments is "yes" by default.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{baseline_interval}{A character value indicating the name of the
interval used as the baseline. Defaults to "t0to5", but can be changed.
Make sure that this matches the baseline interval that you set in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{interval_length}{Length of each interval (in minutes). Used to divide
the dataset into broad ranges for statistical analysis. Important!
\code{max_recording_length} must be evenly divisible by \code{interval_length}.
Defaults to 5.}

\item{parameter}{A character value specifying the parameter to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
parameter is "amplitude", which contains amplitudes normalized relative to
the baseline. For spontaneous currents (\code{current_type = "sEPSC"}), the
available parameters are "amplitude" (normalized currents),
"raw_amplitude", "frequency" (normalized frequency) or "raw_frequency".}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{save_output_as_RDS}{A character ("yes" or "no") describing if the
resulting object should be saved as an RDS file in the raw data folder.}
}
\value{
A dataframe
}
\description{
This function enables you to perform a series of paired t-tests comparing the
mean current amplitude within each interval relative to the mean current
amplitude during the baseline. This uses the \code{paired_t_test} function from
\code{rstatix}, with \code{paired = TRUE} and Holm's adjustment for multiple
comparisons (\code{p.adjust.method = "holm"}). The resulting output table can also
be used to apply significance stars to the plot in \code{make_summary_plot()}.
}
\examples{
perform_t_tests_for_summary_plot(data = sample_summary_eEPSC_df,
 test_category = 2,
 include_all_treatments = "yes",
 treatment_colour_theme = sample_treatment_names_and_colours,
 parameter = "amplitude",
 baseline_interval = "t0to5",
 interval_length = 5,
 list_of_treatments = NULL,
 current_type = "eEPSC",
 save_output_as_RDS = "no")
}
\seealso{
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} for an example of how the normalized
current amplitudes were created.
}
