% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_PPR_data_single_treatment}
\alias{plot_PPR_data_single_treatment}
\title{Make a PPR plot for a single treatment}
\usage{
plot_PPR_data_single_treatment(
  data,
  plot_treatment = "Control",
  plot_category = 2,
  baseline_label = "Baseline",
  post_hormone_label = "Post-hormone",
  test_type,
  large_axis_text = "no",
  geom_signif_size = 0.4,
  treatment_colour_theme,
  theme_options,
  save_plot_png = "no",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{Paired pulse ratio data generated from \code{\link[=make_PPR_data]{make_PPR_data()}}.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control"). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{baseline_label}{A character value for the x-axis label applied to the
pre-hormone state. Defaults to "Baseline".}

\item{post_hormone_label}{A character value for x-axis label applied to
the post-hormone or post-protocol state. Defaults to "Post-hormone" but you
will likely change this to the hormone or protocol name.}

\item{test_type}{A character (must be "wilcox.test", "t.test" or "none")
describing the statistical model used to create a significance bracket
comparing the pre- and post-hormone groups.}

\item{large_axis_text}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which increases the size of the axis text.}

\item{geom_signif_size}{A numeric value describing the size of the geom_signif bracket size. Defaults to 0.4, which is a good thickness for most applications.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"}, it will also generate
a .png file in the folder \code{Figures/Evoked-currents/PPR} relative to the
project directory. The treatment will be included in the filename.
}
\description{
\code{plot_PPR_data_single_treatment()} creates a categorical scatter plot with
experimental state (i.e. baseline/before and after) on the x-axis and the
paired-pulse ratio (PPR) on the y-axis. There are also lines connecting the
"before" data point to the "after" data point for each letter.
}
\details{
If you specify a \code{test_type}, the function will perform a paired t-test or
paired wilcox test and add brackets with significance stars through
\code{ggsignif::geom_signif()}.
}
\examples{
plot_PPR_data_single_treatment(
  data = sample_PPR_df,
  plot_treatment = "Control",
  plot_category = 2,
  baseline_label = "Baseline",
  post_hormone_label = "Insulin",
  test_type = "t.test",
  large_axis_text = "no",
  treatment_colour_theme = sample_treatment_names_and_colours,
  theme_options = sample_theme_options,
  save_plot_png = "no"
)

}
\seealso{
\code{\link[=plot_PPR_data_multiple_treatments]{plot_PPR_data_multiple_treatments()}} to plot changes in PPR for multiple treatments.
}
