% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_spontaneous_current_parameter_comparison}
\alias{plot_spontaneous_current_parameter_comparison}
\title{Visually compare spontaneous current parameters}
\usage{
plot_spontaneous_current_parameter_comparison(
  data,
  plot_category,
  plot_treatment,
  parameter,
  baseline_interval,
  post_hormone_interval,
  large_axis_text = "no",
  hormone_added,
  treatment_colour_theme,
  theme_options,
  save_plot_png
)
}
\arguments{
\item{data}{Summary data for spontaneous currents generated using \code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}} where \code{current_type == "sEPSC"}.}

\item{parameter}{A character value ("raw_amplitude" or "raw_frequency") only.
Normalized amplitude and frequency are not available because all baseline
values are 100.}
}
\value{
A ggplot object
}
\description{
\code{\link[=plot_spontaneous_current_parameter_comparison]{plot_spontaneous_current_parameter_comparison()}} is a useful function to see
how spontaneous current amplitude or frequency change after adding a hormone.
This function produces sina plots (with raw datapoints on top) for the raw
amplitude (or raw frequency) for two time intervals: the baseline interval,
and a user-specified interval after a hormone or other modification has been
applied.
}
\examples{
plot_spontaneous_current_parameter_comparison(
 data = sample_summary_sEPSC_df,
 plot_category = 2,
 plot_treatment = "Control",
 parameter = "raw_amplitude",
 large_axis_text = "no",
 hormone_added = "Insulin",
 baseline_interval = "t0to5",
 post_hormone_interval = "t20to25",
 treatment_colour_theme = sample_treatment_names_and_colours,
 theme_options = sample_theme_options,
 save_plot_png = "no"
)

}
