% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_spontaneous_current_parameter_comparison}
\alias{plot_spontaneous_current_parameter_comparison}
\title{Visually compare spontaneous current parameters}
\usage{
plot_spontaneous_current_parameter_comparison(
  data,
  plot_category = 2,
  plot_treatment = "Control",
  included_sexes = "both",
  facet_by_sex = "no",
  male_label = "Male",
  female_label = "Female",
  left_sex = "Female",
  y_variable = "raw_amplitude",
  y_variable_signif_brackets = NULL,
  hormone_added = "Insulin",
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  test_type,
  map_signif_level_values = F,
  geom_signif_family = "",
  geom_signif_text_size = 5,
  large_axis_text = "no",
  treatment_colour_theme,
  mean_line_thickness = 1.2,
  mean_point_size = 2.5,
  geom_signif_size = 0.5,
  theme_options,
  plot_type = "violin",
  save_plot_png = "no",
  filename_suffix = "",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{Summary data for spontaneous currents generated using \code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}} where \code{current_type == "sEPSC"}.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. \code{"Control"}). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{included_sexes}{A character value (\code{"both"}, \code{"male"} or \code{"female"}). Useful if you want to have a plot with data from one sex only. Defaults to \code{"both"}. If you choose a single sex, the resulting plot will have \code{"-males-only"} or \code{"-females-only"} in the file name. WARNING!! If you choose \code{"male"} or \code{"female"}, you MUST ensure that the \code{t_test_df} contains data that has been filtered to only include one sex. Otherwise, the significance stars will represent both sexes and it will be inaccurate.}

\item{facet_by_sex}{A character value (\code{"yes"} or \code{"no"}) describing if the plots should be faceted by sex. This is only available if \code{"included_sexes"} is \code{"both"}. The resulting plot will be split in two, with male data on the left and female data on the right.}

\item{male_label}{A character value used to describe how males are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for male data in the \code{sex} column, and it must be consistent across data sheets. Defaults to \code{"Male"}.}

\item{female_label}{A character value used to describe how females are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for female data in the \code{sex} column, and it must be consistent across data sheets. This must be consistent in all data sheets. Defaults to \code{"Female"}.}

\item{left_sex}{A character value ("Female" or "Male") describing the sex that will appear on the left side of a faceted plot. Only applies if \code{facet_by_sex} is \code{"yes"}.}

\item{y_variable}{A character value (\code{"raw_amplitude"} or \code{"raw_frequency"}) only.
Normalized amplitude and frequency are not available because all baseline
values are 100.}

\item{y_variable_signif_brackets}{A character value. You should only use this if your data did not pass assumptions and you had to transform it. \code{y_variable_signif_brackets} should be the name of the column of \code{data} which has the transformed data (e.g. log-transformed data). Raw data will be plotted, but the significance brackets (and t-test/wilcox test) will use the transformed data. If you did not transform the data, leave this argument blank, and the function will automatically use the correct column associated with \code{y_variable}.}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
\code{"500 nM Insulin"}, \code{"CCK + Leptin"}, and \code{"Insulin"}. If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{baseline_interval}{A character value indicating the name of the
interval used as the baseline. Defaults to \code{"t0to5"}, but can be changed.
Make sure that this matches the baseline interval that you set in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}.}

\item{post_hormone_interval}{A character value specifying the interval used
for the data points after a hormone or protocol was applied. This must
match the \code{post_hormone_interval} used in \code{\link[=make_variance_data]{make_variance_data()}}.}

\item{test_type}{A character (must be \code{"wilcox.test"}, \code{"t.test"} or \code{"none"})
describing the statistical model used to create a significance bracket
comparing the pre- and post-hormone groups.}

\item{map_signif_level_values}{A \code{TRUE/FALSE} value or a list of character values for mapping p-values. If \code{TRUE}, p-values will be mapped with asterisks (e.g. \* for p < 0.05, for p < 0.01). If \code{FALSE}, raw p-values will display. You can also insert a list of custom mappings or a function. For example, use  \code{map_signif_level_values = function(p) if (p < 0.1) {round(p, 3)} else {"ns"}} to only display the p-values when they are below 0.1.}

\item{geom_signif_family}{A character value describing the font family used for the p-value annotations used by \code{ggsignif::geom_signif()}. Defaults to \code{""} (empty value, will be replaced with default system font), but can be replaced with a named font. Use a package like \code{extrafont} to load system fonts into R.}

\item{geom_signif_text_size}{A numeric value describing the size of the text annotations (significance stars or p-values) on the plot. Defaults to \code{8}.}

\item{large_axis_text}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, a ggplot theme
layer will be applied which increases the axis text.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{mean_line_thickness}{A numeric value describing the thickness of the line used to indicate the mean for a group. Defaults to \code{1.2}.}

\item{mean_point_size}{A numeric value describing the size of the points used to indicate the means. Defaults to \code{2.5}.}

\item{geom_signif_size}{A numeric value describing the size of the \code{geom_signif} bracket size. Defaults to \code{0.5}, which is a good thickness for most applications.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{plot_type}{A character value (\code{"violin"} or \code{"lineplot"}) which determines how the data are plotted. This can be a pale gray violin plot with a sina plot overlay, or a lineplot where "before" and "after" values for each cell are connected with a line.}

\item{save_plot_png}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below \verb{Figures/} in your project directory.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you want to specify anything about the data (for example, to distinguish between recordings produced in MiniAnalysis
vs. Clampfit).}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A ggplot object
}
\description{
\code{\link[=plot_spontaneous_current_parameter_comparison]{plot_spontaneous_current_parameter_comparison()}} is a useful function to see
how spontaneous current amplitude or frequency change after adding a hormone.
This function produces sina plots (with raw datapoints on top) for the raw
amplitude (or raw frequency) for two time intervals: the baseline interval,
and a user-specified interval after a hormone or other modification has been
applied.
}
\details{
If you specify a \code{test_type}, the function will perform a paired t-test or
paired wilcox test and add brackets with significance stars through
\code{ggsignif::geom_signif()}.
}
\examples{
plot_spontaneous_current_parameter_comparison(
  data = sample_summary_sEPSC_df$summary_data,
  plot_category = 2,
  plot_treatment = "Control",
  y_variable = "raw_amplitude",
  included_sexes = "both",
  facet_by_sex = "no",
  hormone_added = "Insulin",
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  test_type = "wilcox.test",
  large_axis_text = "no",
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_plot_png = "no",
  theme_options = sample_theme_options
)

# Facet by sex

# NOTE: To reduce the package file size, there is only one data point
# per sex for the sEPSC data. This is why there are error messages in the example.
# The plot will not have these errors if you have more than one data point
# per sex.

plot_spontaneous_current_parameter_comparison(
  data = sample_summary_sEPSC_df$summary_data,
  plot_category = 2,
  plot_treatment = "Control",
  y_variable = "raw_amplitude",
  included_sexes = "both",
  facet_by_sex = "yes",
  hormone_added = "Insulin",
  baseline_interval = "t0to5",
  post_hormone_interval = "t20to25",
  test_type = "wilcox.test",
  large_axis_text = "no",
  treatment_colour_theme = sample_treatment_names_and_colours,
  save_plot_png = "no",
  theme_options = sample_theme_options
)

}
