% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_summary_current_data}
\alias{plot_summary_current_data}
\title{Make a summary plot for a specific treatment}
\usage{
plot_summary_current_data(
  data,
  plot_category = 2,
  plot_treatment = "Control",
  current_type = "eEPSC",
  y_variable = "amplitude",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  hormone_end_time = NULL,
  legend_position = "right",
  legend_position_inside = c(0.13, 0.15),
  included_sexes = "both",
  male_label = "Male",
  female_label = "Female",
  include_sex_in_legend = "yes",
  include_representative_trace = "no",
  representative_trace_filename = NULL,
  position_dodge_size = 0,
  annotation_x_min = 1,
  annotation_x_max = 8,
  annotation_y_min = 0,
  annotation_y_max = 40,
  y_axis_limit,
  x_label = "Time (min)",
  signif_stars = "no",
  significance_display_method = "stars",
  geom_signif_text_size = 5,
  geom_signif_family = "",
  geom_point_size_large_axis = 1.3,
  geom_point_size = 0.9,
  geom_point_alpha = 1,
  t_test_df = NULL,
  t_test_df_male = NULL,
  t_test_df_female = NULL,
  stars_position = NULL,
  stars_position_male = NULL,
  stars_position_female = NULL,
  stars_colour = "black",
  stars_colour_male = "black",
  stars_colour_female = "black",
  large_axis_text = "no",
  shade_intervals = "no",
  theme_options,
  treatment_colour_theme,
  save_plot_png = "no",
  filename_suffix = "",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{A dataframe containing pruned summary data for all cells. This is
the \verb{$all_cells} element of the list generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. \code{"Control"}). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{current_type}{A character describing the current type. Allowed values
are \code{"eEPSC"} or \code{"sEPSC"}.}

\item{y_variable}{A character value specifying the variable to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
y_variables are \code{"P1"}, \code{"P1_transformed"}, \code{"mean_P1"} and \code{"PPR"}. \emph{Note}: If you
select \code{"mean_P1"}, you must set the \code{pruned} argument to \code{"yes"}. For
spontaneous currents (\code{current_type = "sEPSC"}), the available y_variables
are \code{"amplitude"} or \code{"frequency"}. NOTE: \code{"frequency"} is only available if \code{pruned = "yes"}.}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
\code{"500 nM Insulin"}, \code{"CCK + Leptin"}, and \code{"Insulin"}. If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{hormone_or_HFS_start_time}{A numeric value indicating the time (in
minutes) when a hormone was added or when HFS was applied. This will set
the annotation line start point. This can also be set to \code{NULL} if you want to create your own more complex annotations (for example, for washout, or multiple hormones added at different times). To add your own annotations, use \code{ggplot2::annotate(geom = "segment")} as a starting point.}

\item{hormone_end_time}{A numeric value indicating the time (in minutes) when
a hormone stopped being applied, such as for a washout experiment.}

\item{legend_position}{A character value (\code{"left"}, \code{"right"}, \code{"top"}, \code{"bottom"}, \code{"inside"}) which describes the location of the legend. If you set this to \code{"inside"}, you may need to adjust the location of the representative trace with \code{annotation_x_min}, \code{annotation_x_max}, etc. Or, you can choose to insert the representative trace as a separate geom using the \code{patchwork} package (e.g. set up a two-row patchwork with the representative trace in the top row, and the summary plot in the second row).}

\item{legend_position_inside}{A list of two numeric values describing the coordinates of the legend. This only applies if \code{legend_position = "inside"}.}

\item{included_sexes}{A character value (\code{"both"}, \code{"male"} or \code{"female"}). Useful if you want to have a plot with data from one sex only. Defaults to \code{"both"}. If you choose a single sex, the resulting plot will have \code{"-males-only"} or \code{"-females-only"} in the file name. WARNING!! If you choose \code{"male"} or \code{"female"}, you MUST ensure that the \code{t_test_df} accurately represents what you want to demonstrate: significance for both sexes grouped together (use \code{t_test_df}) or significance for each sex separately (use \code{t_test_df_male} and \code{t_test_df_female} and adjust these heights using \code{stars_position_male} and \code{stars_position_female}).}

\item{male_label}{A character value used to describe how males are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for male data in the \code{sex} column, and it must be consistent across data sheets. Defaults to \code{"Male"}.}

\item{female_label}{A character value used to describe how females are encoded in the \code{sex} column of the dataframe used in \code{data}. This MUST match the value for female data in the \code{sex} column, and it must be consistent across data sheets. This must be consistent in all data sheets. Defaults to \code{"Female"}.}

\item{include_sex_in_legend}{A character (\code{"yes"} or \code{"no"}) describing if the legend should include the sexes or just cell counts. For example, this determines if the legend will be "n = 3" or "Males, n = 3".}

\item{include_representative_trace}{A character (\code{"yes"} or \code{"no"}) describing if
a representative trace should be included as an overlay to the plot. This
pulls from a png file stored in \verb{Figures/Representative-Traces/}". Please
specify the file-name in \code{representative_trace_filename}.}

\item{representative_trace_filename}{A character value describing the
filename of the representative trace. This should be the name of a .png
file. Use relative paths to specify files. For example, a figure in the \code{Figures/Representative-Traces} subfolder would be entered as \code{representative_trace_filename = "Figures/Representative-Traces/Category-2-Control-Trace.png"}.}

\item{position_dodge_size}{A numeric value describing the distance that points should be dodged through \code{ggplot2::position_dodge()}. Defaults to \code{0}.}

\item{annotation_x_min}{A numeric value describing the minimum value on the x-axis for the representative trace. Defaults to \code{1}, which will place it at the lower left corner of the plot (when combined with the default value for \code{annotation_y_min}).}

\item{annotation_x_max}{A numeric value describing the maximum value on the x-axis for the representative trace. Change this if your representative trace image looks squished or stretched.}

\item{annotation_y_min}{A numeric value describing the minimum value on the y-axis for the representative trace. Defaults to \code{0}, which will place it at the lower left corner of the plot (when combined with the default value for \code{annotation_x_min}).}

\item{annotation_y_max}{A numeric value describing the maximum value on the y-axis for the representative trace. Change this if your representative trace image looks squished or stretched.}

\item{y_axis_limit}{A numeric value describing the maximum value on the y-axis.}

\item{x_label}{A character value specifying the x-axis label. Defaults to "Time (min)".}

\item{signif_stars}{A character (\code{"yes"} or \code{"no"}) describing if significance
stars should be included as an overlay in the plot. If \code{"yes"}, you must
specify a dataframe containing the results of a t-test, which is generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. NOTE! If \code{included_sexes} is \code{"male"} or \code{"female"}, you MUST use a t-test that was performed on data filtered to ONE sex.}

\item{significance_display_method}{A character value (\code{"stars"} or \code{"p-value"}) describing how significance values should be displayed. These annotations will not appear if \code{signif_stars} is \code{"no"}.)}

\item{geom_signif_text_size}{A numeric value describing the size of the text annotations (significance stars or p-values) on the plot. Defaults to \code{8}.}

\item{geom_signif_family}{A character value describing the font family used for the p-value annotations used by \code{ggsignif::geom_signif()}. Defaults to \code{""} (empty, will use default system font), but can be replaced with a named font. Use a package like \code{extrafont} to load system fonts into R.}

\item{geom_point_size_large_axis}{A numeric value describing the \code{geom_point} size when you have set \code{large_axis_text = "yes"}. Defaults to 3.5.}

\item{geom_point_size}{A numeric value describing the typical \code{geom_point} size when you do not have \code{large_axis_text} enabled. Defaults to 0.9.}

\item{geom_point_alpha}{A numeric value describing the alpha of \code{geom_point} or \code{geom_pointrange}. Defaults to \code{0.7}.}

\item{t_test_df}{A dataframe of t-test results, which has been generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. Important note! The t-test
dataframe must be filtered to match the same conditions in the \code{data}
argument, or the significance stars will be misleading. For example, if you want to show
significance for both sexes as a group, use this.
If you want to show male and female significance values separately,
use \code{t_test_df_male} and \code{t_test_df_female} and adjust their heights
with \code{stars_position_male} and \code{stars_position_female}. If you only have data for one sex, you could use either option.}

\item{t_test_df_male}{A dataframe of t-test results (male only), which has been generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. This data has been filtered to only include male data before being put through \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}.}

\item{t_test_df_female}{A dataframe of t-test results (female only), which has been generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. This data has been filtered to only include female data before being put through \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}.}

\item{stars_position}{A numeric value stating the position on the y-axis of
the significance stars. If you want to specify these for the sexes separately,
use \code{stars_position_male} and \code{stars_position_female}. Defaults to \code{NULL} which automatically defaults to \code{y_axis_limit - 50}.}

\item{stars_position_male}{A numeric value stating the position on the y-axis of
the significance stars for the male data specified by \code{t_test_df_male}. Defaults to \code{NULL} which sets it at \code{y_axis_limit - 50}.}

\item{stars_position_female}{A numeric value stating the position on the y-axis of
the significance stars for the male data specified by \code{t_test_df_female}. Defaults to \code{NULL} which sets it at \code{y_axis_limit - 75}.}

\item{stars_colour}{A character value (can be a hex code or a named R colour like \code{"red"}) describing the colour of the significance stars. Defaults to \code{"Black"}.}

\item{stars_colour_male}{A character value (can be a hex code or a named R colour like \code{"red"}) describing the colour of the significance stars for male data. Use only if you have put data in \code{t_test_df_male}. Defaults to \code{"Black"}.}

\item{stars_colour_female}{A character value (can be a hex code or a named R colour like \code{"red"}) describing the colour of the significance stars for female data. Use only if you have put data in \code{t_test_df_female}. Defaults to \code{"Black"}.}

\item{large_axis_text}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, a ggplot theme layer will be applied which increases the size of the axis text.}

\item{shade_intervals}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, a ggplot theme
layer will be applied which adds lightly shaded rectangles to highlight
5-minute intervals.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_plot_png}{A character (\code{"yes"} or \code{"no"}). If \code{"yes"}, the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below \verb{Figures/} in your project directory.}

\item{filename_suffix}{Optional character value to add a suffix to the
filename of the .png file created with this plot. Could be useful if you want to specify anything about the data (for example, to distinguish between recordings produced in MiniAnalysis
vs. Clampfit).}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"}, it will also generate
a .png file exported to \code{Figures/Evoked-currents/Output-summary-plots} or
\code{Figures/Spontaneous-currents/Output-summary-plots}, depending on the
\code{current_type}. The .png filename will be in this format:
"Summary-plot-\code{plot_treatment}-category-\code{plot_category}-\code{file_name_ending}-\code{text_size}-\code{sexes}.png".
The \code{text_size} will only be added on if you are using \code{large_axis_text}
("LARGE" will be included in the filename). The \code{file_name_ending} will be
automatically added on for spontaneous current data to specify what
y_variable is plotted (e.g. \code{"raw_amplitude"}). The \code{sexes} will only be added if you choose to plot a single sex. For example, if you set \code{included_sexes = "Male"}, the \code{.png} filename will have \code{"-males-only"} included in the filename.

Example auto-generated filenames include:
\itemize{
\item \code{"Summary-plot-Control-category-2.png"}
\item \code{"Summary-plot-Control-category-2-males-only.png"}
\item \code{"Summary-plot-Control-category-2_raw_amplitude.png"} for spontaneous
currents.
}
}
\description{
This function enables you to create a scatterplot of mean evoked (or
spontaneous) current amplitude over over time in minutes. The data are
summarized by treatment and sex, and averaged per minute. The data are
presented as mean +/- the standard error.
}
\examples{

# Both sexes with significance stars for separate sexes
# Note: The significance stars are sample data


plot_summary_current_data(
  data = sample_pruned_eEPSC_df$all_cells,
  plot_category = 2,
  plot_treatment = "Control",
  current_type = "eEPSC",
  y_variable = "amplitude",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  included_sexes = "both",
  include_representative_trace = "yes",
  representative_trace_filename = import_ext_data("Control-trace.png"),
  y_axis_limit = 175,
  signif_stars = "yes",
  t_test_df_male = sample_eEPSC_t_test_df_male,
  t_test_df_female = sample_eEPSC_t_test_df_female,
  stars_position_male = 30,
  stars_position_female = 70,
  stars_colour_male = "#6600cc",
  stars_colour_female = "#d6b8f5",
  large_axis_text = "no",
  shade_intervals = "no",
  treatment_colour_theme = sample_treatment_names_and_colours,
  theme_options = sample_theme_options
)
# Both sexes with significance stars from grouped data

plot_summary_current_data(
  data = sample_pruned_eEPSC_df$all_cells,
  plot_category = 2,
  plot_treatment = "Control",
  current_type = "eEPSC",
  y_variable = "amplitude",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  hormone_end_time = 25,
  included_sexes = "both",
  include_representative_trace = "yes",
  representative_trace_filename = import_ext_data("Control-trace.png"),
  y_axis_limit = 175,
  signif_stars = "yes",
  t_test_df = sample_eEPSC_t_test_df,
  large_axis_text = "no",
  shade_intervals = "no",
  treatment_colour_theme = sample_treatment_names_and_colours,
  theme_options = sample_theme_options
)
}
\seealso{
\code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}} which produces the significance
stars appended to the plot.

\code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}} for the function that will produce the
summary data used in this plot.

\code{\link[=plot_change_as_connected_lines]{plot_change_as_connected_lines()}} to show this same data as lines connecting the "before" and "after" points.
}
