% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_summary_current_data}
\alias{plot_summary_current_data}
\title{Make a summary plot for a specific treatment}
\usage{
plot_summary_current_data(
  data,
  plot_category = 2,
  plot_treatment = "Control",
  current_type = "eEPSC",
  y_variable = "amplitude",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  included_sexes = "both",
  male_label = "Male",
  female_label = "Female",
  include_representative_trace = "no",
  representative_trace_filename = NULL,
  annotation_x_min = 1,
  annotation_x_max = 8,
  annotation_y_min = 0,
  annotation_y_max = 40,
  y_axis_limit,
  signif_stars = "no",
  significance_display_method = "stars",
  geom_signif_size = 8,
  geom_signif_family = "",
  t_test_df,
  large_axis_text = "no",
  shade_intervals = "no",
  theme_options,
  treatment_colour_theme,
  save_plot_png = "no",
  ggplot_theme = patchclampplotteR_theme()
)
}
\arguments{
\item{data}{A dataframe containing pruned summary data for all cells. This is
the third element of the list generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control"). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{y_variable}{A character value specifying the variable to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
y_variables are "P1", "P1_transformed", "mean_P1" and "PPR". \emph{Note}: If you
select "mean_P1", you must set the \code{pruned} argument to "yes". For
spontaneous currents (\code{current_type = "sEPSC"}), the available y_variables
are "amplitude" or "frequency".}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
"500 nM Insulin", "CCK + Leptin", and "Insulin". If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{hormone_or_HFS_start_time}{A numeric value indicating the time (in
minutes) when a hormone was added or when HFS was applied. This will set
the annotation line start point.}

\item{included_sexes}{A character value ("both", "male" or "female"). Useful if you want to have a plot with data from one sex only. Defaults to "both". If you choose a single sex, the resulting plot will have "-males-only" or "-females-only" in the file name. WARNING!! If you choose "male" or "female", you MUST ensure that the \code{t_test_df} contains data that has been filtered to only include one sex. Otherwise, the significance stars will represent both sexes and it will be inaccurate.}

\item{male_label}{A character value used to describe how males are encoded in the \code{sex} column of the dataframe used in \code{data}. Examples include "Males", "Male", "male", "males", "M", etc. Defaults to "Male".}

\item{female_label}{A character value used to describe how females are encoded in the \code{sex} column of the dataframe used in \code{data}. Examples include "Females", "Female", "female", "females", "F", etc. Defaults to "Female".}

\item{include_representative_trace}{A character ("yes" or "no") describing if
a representative trace should be included as an overlay to the plot. This
pulls from a png file stored in \verb{Figures/Representative-Traces/}". Please
specify the file-name in \code{representative_trace_filename}.}

\item{representative_trace_filename}{A character value describing the
filename of the representative trace. This should be the name of a .png
file. Use relative paths to specify files. For example, a figure in the Figures/Representative-Traces subfolder would be entered as \code{representative_trace_filename = "Figures/Representative-Traces/Category-2-Control-Trace.png"}.}

\item{annotation_x_min}{A numeric value describing the minimum value on the x-axis for the representative trace. Defaults to 1, which will place it at the lower left corner of the plot (when combined with the default value for \code{annotation_y_min}).}

\item{annotation_x_max}{A numeric value describing the maximum value on the x-axis for the representative trace. Change this if your representative trace image looks squished or stretched.}

\item{annotation_y_min}{A numeric value describing the minimum value on the y-axis for the representative trace. Defaults to 0, which will place it at the lower left corner of the plot (when combined with the default value for \code{annotation_x_min}).}

\item{annotation_y_max}{A numeric value describing the maximum value on the y-axis for the representative trace. Change this if your representative trace image looks squished or stretched.}

\item{y_axis_limit}{A numeric value describing the maximum value on the y-axis.}

\item{signif_stars}{A character ("yes" or "no") describing if significance
stars should be included as an overlay in the plot. If "yes", you must
specify a dataframe containing the results of a t-test, which is generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. NOTE! If \code{included_sexes} is "male" or "female", you MUST use a t-test that was performed on data filtered to ONE sex.}

\item{significance_display_method}{A character value ("stars" or "p-value") describing how significance values should be displayed. These annotations will not appear if \code{signif_stars} is "no".)}

\item{geom_signif_size}{A numeric value describing the size of the text annotations (significance stars or p-values) on the plot. Defaults to \code{8}.}

\item{geom_signif_family}{A character value describing the font family used for the p-value annotations used by \code{ggsignif::geom_signif()}.}

\item{t_test_df}{A dataframe of t-test results, which has been generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. Important note! The t-test
dataframe must be filtered to match the same conditions in the \code{data}
argument, or the significance stars will be misleading.}

\item{large_axis_text}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which increases the size of the axis text.}

\item{shade_intervals}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which adds lightly shaded rectangles to highlight
5-minute intervals.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}

\item{ggplot_theme}{The name of a ggplot theme or your custom theme. This will be added as a layer to a ggplot object. The default is \code{patchclampplotteR_theme()}, but other valid entries include \code{theme_bw()}, \code{theme_classic()} or the name of a custom ggplot theme stored as an object.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"}, it will also generate
a .png file exported to \code{Figures/Evoked-currents/Output-summary-plots} or
\code{Figures/Spontaneous-currents/Output-summary-plots}, depending on the
\code{current_type}. The .png filename will be in this format:
"Summary-plot-\code{plot_treatment}-category-\code{plot_category}-\code{file_name_ending}-\code{text_size}-\code{sexes}.png".
The \code{text_size} will only be added on if you are using \code{large_axis_text}
("LARGE" will be included in the filename). The \code{file_name_ending} will be
automatically added on for spontaneous current data to specify what
y_variable is plotted (e.g. "raw_amplitude"). The \code{sexes} will only be added if you choose to plot a single sex. For example, if you set \code{included_sexes = "Male"}, the \code{.png} filename will have "-males-only" included in the filename.

Example auto-generated filenames include:
\itemize{
\item "Summary-plot-Control-category-2.png"
\item "Summary-plot-Control-category-2-males-only.png"
\item "Summary-plot-Control-category-2_raw_amplitude.png" for spontaneous
currents.
}
}
\description{
This function enables you to create a scatterplot of mean evoked (or
spontaneous) current amplitude over over time in minutes. The data are
summarized by treatment and sex, and averaged per minute. The data are
presented as mean +/- the standard error.
}
\examples{

plot_summary_current_data(
  data = sample_pruned_eEPSC_df$all_cells,
  plot_category = 2,
  plot_treatment = "Control",
  current_type = "eEPSC",
  y_variable = "amplitude",
  hormone_added = "Insulin",
  hormone_or_HFS_start_time = 5,
  include_representative_trace = "yes",
  representative_trace_filename = import_ext_data("Control-trace.png"),
  y_axis_limit = 175,
  signif_stars = "yes",
  t_test_df = sample_eEPSC_t_test_df,
  large_axis_text = "no",
  shade_intervals = "no",
  treatment_colour_theme = sample_treatment_names_and_colours,
  theme_options = sample_theme_options
)
}
\seealso{
\code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}} which produces the significance
stars appended to the plot.

\code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}} for the function that will produce the
summary data used in this plot.
}
