% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-data.R
\name{plot_summary_current_data}
\alias{plot_summary_current_data}
\title{Make a summary plot for a specific treatment}
\usage{
plot_summary_current_data(
  plot_category,
  plot_treatment,
  data,
  current_type,
  parameter,
  hormone_added,
  hormone_or_HFS_start_time,
  include_representative_trace = "no",
  representative_trace_filename,
  t_test_df,
  signif_stars = "no",
  large_axis_text = "no",
  shade_intervals = "no",
  theme_options,
  treatment_colour_theme,
  save_plot_png = "no"
)
}
\arguments{
\item{plot_category}{A numeric value specifying the category, which can be
used to differentiate different protocol types. In the sample dataset for
this package, \code{plot_category == 2} represents experiments where insulin was
applied continuously after a 5-minute baseline period.}

\item{plot_treatment}{A character value specifying the treatment you would
like to plot (e.g. "Control"). \code{plot_treatment} represents antagonists that
were present on the brain slice, or the animals were fasted, etc.}

\item{data}{A dataframe containing pruned summary data for all cells. This is the third element of the list generated from \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}.}

\item{current_type}{A character describing the current type. Allowed values
are "eEPSC" or "sEPSC".}

\item{parameter}{A character value specifying the parameter to be plotted on
the y-axis. For evoked currents (\code{current_type = "eEPSC"}), the available
parameters are "P1", "mean_P1" and "PPR". \emph{Note}: If you select "mean_P1",
you must set the \code{pruned} argument to "yes". For spontaneous currents
(\code{current_type = "sEPSC"}), the available parameters are "amplitude" or
"frequency".}

\item{hormone_added}{A character value that will be used as the label over
the line annotating the period when a hormone was applied. Examples include
"500 nM Insulin", "CCK + Leptin", and "Insulin". If you applied a
high-frequency stimulation (HFS) protocol instead, write "HFS", and an
annotation arrow will be added instead.}

\item{hormone_or_HFS_start_time}{A numeric value indicating the time (in
minutes) when a hormone was added or when HFS was applied. This will set
the annotation line start point.}

\item{include_representative_trace}{A character ("yes" or "no") describing if
a representative trace should be included as an overlay to the plot. This
pulls from a png file stored in \verb{Figures/Representative-Traces/}". Please
specify the file-name in \code{representative_trace_filename}.}

\item{representative_trace_filename}{A character value describing the
filename of the representative trace. This should be the name of a .png
file stored within the subfolder \verb{Figures/Representative-Traces/}.}

\item{t_test_df}{A dataframe of t-test results, which has been generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}. Important note! The t-test
dataframe must be filtered to match the same conditions in the \code{data}
argument, or the significance stars will be misleading.}

\item{signif_stars}{A character ("yes" or "no") describing if significance
stars should be included as an overlay in the plot. If "yes", you must
specify a dataframe containing the results of a t-test, which is generated
using \code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}}.}

\item{large_axis_text}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which increases the size of the axis text.}

\item{shade_intervals}{A character ("yes" or "no"). If "yes", a ggplot theme
layer will be applied which adds lightly shaded rectangles to highlight
5-minute intervals.}

\item{theme_options}{A dataframe containing theme options. See
\link{sample_theme_options} for an example of what this dataframe should
look like.}

\item{treatment_colour_theme}{A dataframe containing treatment names and
their associated colours as hex values. See
\link{sample_treatment_names_and_colours} for an example of what this dataframe
should look like.}

\item{save_plot_png}{A character ("yes" or "no"). If "yes", the plot will be
saved as a .png using ggsave. The filepath depends on the current type, but
they will all go in subfolders below Figures/ in your project directory.}
}
\value{
A ggplot object. If \code{save_plot_png == "yes"}, it will also generate a
.png file exported to \code{Figures/Evoked-currents/Output-summary-plots} or
\code{Figures/Spontaneous-currents/Output-summary-plots}, depending on the
\code{current_type}. The .png filename will be in this format:
"Summary-plot-\code{plot_treatment}-category-\code{plot_category}-\code{file_name_ending}-\code{text_size}.png".
The \code{file_name_ending} and \code{text_size} will only be added on if you are
using a An example filename is: "Summary-plot-Control-category-2.png" or
"Summary-plot-Control-category-2_raw_amplitude.png" for spontaneous
currents.
}
\description{
This function enables you to create a scatterplot of mean evoked (or
spontaneous) current amplitude over over time in minutes. The data are
summarized by treatment and sex, and averaged per minute. The data are
presented as mean +/- the standard error.
}
\examples{

plot_summary_current_data(
  plot_category = 2,
  plot_treatment = "Control",
  data = sample_pruned_eEPSC_df$all_cells,
  current_type = "eEPSC",
  parameter = "amplitude",
  include_representative_trace = "no",
  signif_stars = "yes",
  t_test_df = sample_eEPSC_t_test_df,
  hormone_added = "Insulin",
  large_axis_text = "no",
  shade_intervals = "no",
  hormone_or_HFS_start_time = 5,
  treatment_colour_theme = sample_treatment_names_and_colours,
  theme_options = sample_theme_options
)
}
\seealso{
\code{\link[=perform_t_tests_for_summary_plot]{perform_t_tests_for_summary_plot()}} which produces the significance
stars appended to the plot.

\code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}} for the function that will produce the
summary data used in this plot.
}
