% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_abf}
\alias{sample_abf}
\title{Sample ABF file}
\format{
data
\describe{
\item{episode}{A factor representing the sweep, such as "epi1".}
\item{time_sec}{Time in seconds.}
\item{time}{Time in ms*100.}
\item{voltage}{Numeric value representing voltage in mV.}
\item{current}{Numeric value representing current amplitude in pA.}
}
}
\description{
Sample ABF file
}
\keyword{data}
