% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_ap_abf_insulin}
\alias{sample_ap_abf_insulin}
\title{Action potential recording with insulin}
\format{
A dataframe of 100000 objects of 5 variables.
\describe{
\item{episode}{A factor representing the sweep, such as "epi1".}
\item{time_sec}{Time in seconds.}
\item{time}{Time in ms*100.}
\item{voltage}{Numeric value representing voltage in mV.}
\item{current}{Numeric value representing current amplitude in pA.}
}
}
\description{
This is an excerpt of a raw recording of action potentials imported as a dataframe using \code{\link[=import_ABF_file]{import_ABF_file()}}. It is used to demonstrate the
\code{\link[=plot_AP_trace]{plot_AP_trace()}} function.
}
\details{
This is a current injection protocol taken during after 25 minutes of exposure to 500 nM insulin. See \code{sample_ap_abf_baseline} for data from the same protocol taken during the baseline period.
}
\examples{
plot_AP_trace(
  file = sample_ap_abf_insulin,
  sweeps = c("epi1", "epi10"),
  trace_color = "orange",
  plot_category = 2,
  plot_treatment = "Control",
  state = "Baseline"
)
}
\keyword{data}
