% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_cell_characteristics}
\alias{sample_cell_characteristics}
\title{Information about sex, age, treatment, and animal ID}
\format{
A dataframe with 19 rows and 11 columns
\describe{
\item{letter}{A character value that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics. Example: "A"}
\item{cell}{A character or numeric value representing the cell. For example,
use \verb{3.1.1} for animal #3, slice #1, cell #1.}
\item{sex}{A character value such as "Male" or "Female".}
\item{X}{A numeric value representing the x-value of the cell's location in
µm.}
\item{Y}{A numeric value representing the y-value of the cell's location in
µm.}
\item{age}{A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages). Do not use characters (e.g. avoid
"P31" and use 31 instead).}
\item{animal}{A numeric value representing the animal's ID or number.}
\item{synapses}{A character value such as "Glutamate" or "GABA".}
\item{treatment}{A character value such as "Control" or "HNMPA".}
\item{category}{A numeric value representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as
subgroups. For example, "1" may refer to an experiment where you applied high-frequency stimulation (HFS) to a cell, while "2" is an experiment where you added a hormone like leptin. "3" may be an experiment where you applied HFS in the continuous presence of leptin.}
\item{R_a}{A list of values for the access resistance, which would have been
monitored at several timepoints throughout the recording. See the section
\code{R_a} formatting below.}
}
}
\description{
This dataset provides an example of the type of cell characteristics that you
should be recording for your data. The dataset contains columns for the animal
number, age, sex, the synapses being recorded from, and treatments applied.
There is also a column for access, which is stored as a list.

The most important column is \code{letter}. This is a unique identifier that
you will assign to each recording. The \code{letter} column will enable you to
link all information relevant to the recording (evoked current data,
spontaneous current data, animal data like age, sex, etc.) from different
files.

You should import this .csv file using the \code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}}
function, which will format it in a format that makes it easy to merge with
raw recording data in functions like \code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}. This will
enable you to analyze relationships between properties like age and current
amplitude.
}
\examples{
utils::read.csv(import_ext_data("sample_cell_characteristics.csv"))

}
\seealso{
\code{\link[=import_cell_characteristics_df]{import_cell_characteristics_df()}} to import data like this from a
.csv file.

\code{\link[=make_interactive_summary_table]{make_interactive_summary_table()}} for a function which merges cell
characteristics information with pruned evoked and spontaneous current data
to create an interactive overview table of all recordings.
}
\keyword{data}
