% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_summary_eEPSC_df}
\alias{sample_summary_eEPSC_df}
\title{Evoked current data summarized into 5-minute intervals}
\format{
An .rda file containing a list of three dataframes. The first dataframe (\verb{$percent_change_data}) contains the mean current amplitude for each interval, with a final column (\code{percent_change}) containing the final percent change in amplitude in the last interval relative to the baseline interval. The second dataframe (\verb{$summary_data}) contains summary statistics for each interval. The third dataframe (\verb{$mean_SE}) contains averaged evoked current amplitudes (with standard error) for each time interval. This is useful for publications where you want to report your findings as means +/- SE or SD.

See below for more details.
\subsection{\verb{$percent_change_data}}{

\describe{
\item{\code{category}}{A factor representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups. For example, "1" may refer to an experiment where you applied high-frequency stimulation (HFS) to a cell, while "2" is an experiment where you added a hormone like leptin. "3" may be an experiment where you applied HFS in the continuous presence of leptin.}
\item{\code{letter}}{A factor that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.}
\item{\code{sex}}{A factor with two levels (e.g. "Male" or "Female").}
\item{\code{treatment}}{A factor based on the treatment applied (e.g. "Control",
"HNMPA"). This represents the antagonists or agonists applied, or any
protocol applied to the animals (e.g. "Fasting").}
\item{\code{age}}{A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages).}
\item{\code{animal}}{A numeric value representing the animal's ID or number.}
\item{\code{X}}{A numeric value representing the x-value of the cell's location in
µm. Leave this blank if you don't have this data.}
\item{\code{Y}}{A numeric value representing the y-value of the cell's location in
µm. Leave this blank if you don't have this data.}
\item{\code{synapses}}{A character value (e.g. "Glutamate" or "GABA").}
\item{\code{days_alone}}{A numeric value describing the number of days that the animal was left alone in a cage. This typically ranges from 0 to 2. Fasted animals will have 1 day alone.}
\item{\code{animal_or_slicing_problems}}{A character value ("yes" or "no") describing if there were any issues with the animal (for example, the animal was unusually anxious) or slicing (there were delays during the process, the slices were crumpling, etc.).}
\item{\code{t0to5}}{The mean evoked current amplitude (pA) for this cell during the period of 0 to 5 minutes.}
\item{\code{t5to10}}{The mean evoked current amplitude (pA) for this cell during the period of 5 to 10 minutes.}
\item{\code{t10to15}, \code{t15to20}, \code{tXtY} etc...}{The mean evoked current amplitude (pA) for this cell during the period of \emph{X} to \emph{Y} minutes.}
\item{\code{percent_change}}{The percent change in evoked current amplitude in the interval \code{t20to25} as a percentage of the mean baseline amplitude (\code{t0to5}). For example, if currents began at 100 pA during the baseline period, but were 50 pA by \code{t20to25}, the value of \code{percent_change} will be \verb{50\%} or \code{0.50}.}
}
}

\subsection{\verb{$summary_data}}{

\describe{
\item{\code{category}}{A factor representing the experiment type. Used to
assign top-level groups for further analyses, with \code{treatment} as subgroups. For example, "1" may refer to an experiment where you applied high-frequency stimulation (HFS) to a cell, while "2" is an experiment where you added a hormone like leptin. "3" may be an experiment where you applied HFS in the continuous presence of leptin.}
\item{\code{letter}}{A factor that is a unique identifier for a single
recording. Used to link data sets for evoked or spontaneous currents and
cell-characteristics.}
\item{\code{sex}}{A factor with two levels (e.g. "Male" or "Female").}
\item{\code{treatment}}{A factor based on the treatment applied (e.g. "Control",
"HNMPA"). This represents the antagonists or agonists applied, or any
protocol applied to the animals (e.g. "Fasting").}
\item{\code{interval}}{A factor indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.}
\item{\code{mean_P1_transformed}}{A numeric value representing the mean current
amplitude (in pA normalized to the baseline amplitude) of the first evoked
current. The values used to produce mean_P1_transformed come from all data
points within each interval (the length of the interval was specified in
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}}).}
\item{\code{mean_P1_raw}}{The same values as those in \code{mean_P1_transformed} except
these values contain raw values for current amplitude (e.g. the data were
not baseline transformed).}
\item{\code{n}}{The number of data points used to produce each value in
\code{mean_P1_transformed} and \code{mean_P1_raw}}
\item{\code{sd}}{The standard deviation of the mean current values.}
\item{\code{cv}}{The coefficient of variation of the mean current values.}
\item{\code{se}}{The standard error of the mean current values.}
\item{\code{cv_inverse_square}}{The inverse coefficient of variation squared of
the mean current values. Calculated by using (\code{1/cv^2}), where \code{cv} is the
coefficient of variation. Useful for variance analysis.}
\item{\code{variance}}{The variance of the current values, calculated using \code{stats::var()}.}
\item{\code{VMR}}{The variance-to-mean ratio, calculated using
\code{variance/mean_P1_transformed}. Useful for variance analysis.}
\item{\code{age}}{A numeric value representing the animal's age. Can be any value
as long as the time units are consistent throughout (e.g. don't mix up days
and months when reporting animal ages).}
\item{\code{animal}}{A numeric value representing the animal's ID or number.}
\item{\code{X}}{A numeric value representing the x-value of the cell's location in
µm. Leave this blank if you don't have this data.}
\item{\code{Y}}{A numeric value representing the y-value of the cell's location in
µm. Leave this blank if you don't have this data.}
\item{\code{time}}{A numeric value that represents time in minutes. This column is
autogenerated in \code{\link[=add_new_cells]{add_new_cells()}}.}
\item{\code{synapses}}{A character value (e.g. "Glutamate" or "GABA").}
\item{\code{days_alone}}{A numeric value describing the number of days that the animal was left alone in a cage. This typically ranges from 0 to 2. Fasted animals will have 1 day alone.}
\item{\code{animal_or_slicing_problems}}{A character value ("yes" or "no") describing if there were any issues with the animal (for example, the animal was unusually anxious) or slicing (there were delays during the process, the slices were crumpling, etc.).}
}
}

\subsection{\verb{$mean_SE}}{

\itemize{
\item \code{category} The experiment category (please see \code{import_cell_characteristics_df()} for more details).
\item \code{sex} The sex of the animal
\item \code{treatment} The treatment applied.
\item \code{n} The number of data points (i.e. cells)
\item \code{mean_baseline_raw_P1} The average evoked current amplitude (taken from \code{mean_P1_raw}) during the \code{baseline_interval}.
\item \code{sd_baseline_raw_P1} The standard deviation of \code{mean_baseline_raw_P1.}
\item \code{se_baseline_raw_P1} The standard error of \code{mean_baseline_raw_P1}. Taken by dividing \code{sd_baseline_raw_P1} by the square root of \code{n}.
\item \code{mean_ending_raw_P1} The average evoked current amplitude (taken from \code{mean_P1_raw}) during the \code{ending_interval}.
\item \code{sd_ending_raw_P1} The standard deviation of \code{mean_ending_raw_P1.}
\item \code{se_ending_raw_P1} The standard error of \code{mean_ending_raw_P1}. Taken by dividing \code{sd_ending_raw_P1} by the square root of \code{n}.

\item \code{VMR} The variance-to-mean ratio (VMR) of \code{P1_transformed}.
\item \code{interval} A character value indicating the interval that the data
point belongs to. For example, \code{interval} will be "t0to5" for any data
points from 0 to 5 minutes. Example values: "t0to5", "t5to10", etc.
\item \verb{letter, synapses, sex, treatment, etc.} Unmodified columns from the
original dataset describing the cell's properties.
}
}
}
\description{
This is an example of the summary eEPSC data produced using
\code{\link[=make_summary_EPSC_data]{make_summary_EPSC_data()}}. It is useful for demonstrating functions that
build off of this dataset, such as \code{\link[=plot_summary_current_data]{plot_summary_current_data()}}. This is a list of three dataframes.
}
\seealso{
\code{\link[=make_normalized_EPSC_data]{make_normalized_EPSC_data()}} and \code{\link[=make_pruned_EPSC_data]{make_pruned_EPSC_data()}}
}
\keyword{data}
