% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_theme_options}
\alias{sample_theme_options}
\title{Sample theme options}
\format{
An .rda file containing 23 objects of 2 variables
\describe{
\itemize{
\item \code{option} A character value specifying the treatment applied (e.g.
"Control", "HNMPA").
\item \code{value} A character value with the same treatment names as in
\code{treatment} except you would replace underscores with spaces and add line
breaks if needed. This is to create more attractive, human-readable axis
labels and table values.
}
}
}
\description{
This is an example of the dataframe used to specify the colour themes and plot options (like line thickness).
}
\keyword{data}
