% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_theme_options}
\alias{sample_theme_options}
\title{Sample theme options}
\format{
An .rda file containing 23 objects of 2 variables
\describe{
\itemize{
\item \code{gray_shading_colour} Hex code for the colour used to fill the violin
plots
\item \code{line_col}
\item \code{rectangle_shading_colour} Hex code for the colour used to shade the
rectangles
\item \code{plot_font_family}
\item \code{plot_light_font_family}
\item \code{significance_stars_font}
\item \code{male_shape} Numeric value representing the point shape used for data
values where \code{Sex == Male}
\item \code{female_shape} Numeric value representing the point shape used for
data values where \code{Sex == Female}
\item \code{mean_point_colour}
\item \code{baseline_group_colour}
\item \code{post_hormone_group_colour}
\item \code{connecting_line_width}
\item \code{connecting_line_colour_aps}
\item \code{connecting_line_width_PPR}
\item \code{mean_point_size}
\item \code{geom_sina_size}
\item \code{geom_signif_text_size}
\item \code{AP_trace_size}
\item \code{scale_bar_shift_y}
\item \code{scale_bar_shift_x}
\item \code{y_axis_limit}
\item \code{PPR_y_min}
\item \code{PPR_y_max}
"Control", "HNMPA").
\item \code{line_col} A character value with the same treatment names as in
\code{treatment} except you would replace underscores with spaces and add line
breaks if needed. This is to create more attractive, human-readable axis
labels and table values.
}
}
}
\description{
This is an example of the dataframe used to specify the colour themes and
plot options (like line thickness).
}
\keyword{data}
