% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_theme_options}
\alias{sample_theme_options}
\title{A dataframe of theme options for things like colours and line widths}
\format{
An .rda file containing 23 objects of 2 variables

\describe{
\item{\code{gray_shading_colour}}{Hex code for the colour used to fill the violin
plots.}
\item{\code{line_col}}{Hex code for the colour used for the line indicating
periods when a hormone was applied.}
\item{\code{rectangle_shading_colour}}{Hex code for the colour used to shade the
rectangles.}
\item{\code{male_shape}}{Numeric value representing the point shape used for data
values where \code{Sex == Male}.}
\item{\code{female_shape}}{Numeric value representing the point shape used for
data values where \code{Sex == Female}.}
\item{\code{mean_point_colour}}{Hex code for the colour used to fill the mean data point in plots like the PPR comparison plot.}
\item{\code{connecting_line_width}}{The width of the connecting line in action potential comparison plots.}
\item{\code{connecting_line_colour}}{The colour of the connecting line in action potential comparison plots.}
\item{\code{connecting_line_width_PPR}}{The width of the connecting line in paired-pulse ratio comparison plots.}
\item{\code{mean_point_size}}{The size of the mean data point in plots like the PPR comparison plot.}
\item{\code{geom_signif_text_size}}{The size of the asterisk text in plots like the spontaneous current comparison plots.}
\item{\code{AP_trace_size}}{DO NOT USE; this will be moved to a function argument in future releases.}
\item{\code{y_axis_limit}}{The maximum y-axis value to be used in plots.}
}
}
\description{
This is an example of the dataframe used to specify the colour themes and
plot options (like line thickness). If you want to modify this file, see the section on \strong{Defining your own theme options} below.
}
\section{Defining your own theme options}{


Step 1: Create a .csv file with two columns (\code{option} and \code{value}) modelled after this sample dataset. \emph{Important}: Your .csv must have identical columns and rows as the sample data, or else some plots won't work!

Step 2: Read in the .csv file in using \code{utils::read_csv()}. This will now be an object in your R environment.

Step 3: \strong{Important!!} You must convert the first column (\code{option}) into the \code{rownames}. This is a mandatory step to allow the theme_options to be indexed by row name in plotting functions.

Step 4: Run the following code:

\code{library(tibble)}
\code{my_sample_theme_options <- read.csv(here::here("Data/your_sample_theme_options.csv")) \%>\% remove_rownames \%>\% column_to_rownames(var="option")}

Step 5: Check the resulting object. You should now have 11 objects of 1 variable, and the row names should be \code{gray_shading_colour}, \code{line_col}, etc.

Step 6: Go to a plotting function like \code{plot_raw_current_data()} and replace \code{sample_theme_options} with your newly created object from Step 5.
}

\keyword{data}
