% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-datasets.R
\docType{data}
\name{sample_treatment_names_and_colours}
\alias{sample_treatment_names_and_colours}
\title{A dataframe of treatments and their assigned colours for consistency across plots}
\format{
An .rda file containing 4 objects of 4 variables
\describe{
\item{\code{treatment}}{A character value specifying the treatment applied (e.g.
"Control", "HNMPA") such as antagonists or agonists, or any protocol applied
to the animals (e.g. "Fasting").}
\item{\code{display_names}}{A character value with the same treatment names as in
\code{treatment} except you would replace underscores with spaces and add line
breaks if needed. This is to create more attractive, human-readable axis
labels and table values.}
\item{\code{colours}}{A character value with hex colour values.}
\item{\code{very_pale_colours}}{A character value with hex colour values. These
are all lighter than the colours in \code{colours} and they are used to create
contrast in summary plots with both sexes.}
}
}
\description{
This is an example of the dataframe used to specify a colour theme for each
treatment. It is used to create a consistent colour scheme across all plots
produced with this package.
}
\keyword{data}
